Refactor MainWindow into modular mixins
