+++ codes/mainwindow.py
@@ -53,6 +53,8 @@ from gui.main_window.theme_mixin import ThemeMixin
 from gui.main_window.ga_mixin import GAOptimizationMixin
 from gui.main_window.frf_mixin import FRFMixin
 from gui.main_window.pso_mixin import PSOMixin
+from gui.main_window.input_mixin import InputTabsMixin
+from gui.main_window.extra_opt_mixin import ExtraOptimizationMixin
 
 
 # Additional libraries used
@@ -83,7 +85,9 @@ except ImportError:
 
         
 
-class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin, ThemeMixin, FRFMixin, PSOMixin, GAOptimizationMixin):
+class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin,
+                 ThemeMixin, FRFMixin, PSOMixin, GAOptimizationMixin,
+                 InputTabsMixin, ExtraOptimizationMixin):
     def __init__(self):
         super().__init__()
         self.setWindowTitle("DeVana")
@@ -541,654 +545,6 @@ class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin
         
         # Add page to content stack
         self.content_stack.addWidget(stochastic_page)
-
-    def create_main_system_tab(self):
-        """Create the main system parameters tab"""
-        self.main_system_tab = QWidget()
-        layout = QVBoxLayout(self.main_system_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Main system parameters group
-        params_group = QGroupBox("Main System Parameters")
-        params_layout = QFormLayout(params_group)
-        
-        # MU parameter
-        self.mu_box = QDoubleSpinBox()
-        self.mu_box.setRange(-1e6, 1e6)
-        self.mu_box.setDecimals(6)
-        self.mu_box.setValue(1.0)
-        params_layout.addRow("μ (MU):", self.mu_box)
-
-        # LANDA parameters (Lambda)
-        self.landa_boxes = []
-        for i in range(5):
-            box = QDoubleSpinBox()
-            box.setRange(-1e6, 1e6)
-            box.setDecimals(6)
-            if i < 2:
-                box.setValue(1.0)
-            else:
-                box.setValue(0.5)
-            self.landa_boxes.append(box)
-            params_layout.addRow(f"Λ_{i+1}:", box)
-
-        # NU parameters
-        self.nu_boxes = []
-        for i in range(5):
-            box = QDoubleSpinBox()
-            box.setRange(-1e6, 1e6)
-            box.setDecimals(6)
-            box.setValue(0.75)
-            self.nu_boxes.append(box)
-            params_layout.addRow(f"Ν_{i+1}:", box)
-
-        # A_LOW parameter
-        self.a_low_box = QDoubleSpinBox()
-        self.a_low_box.setRange(0, 1e10)
-        self.a_low_box.setDecimals(6)
-        self.a_low_box.setValue(0.05)
-        params_layout.addRow("A_LOW:", self.a_low_box)
-
-        # A_UPP parameter
-        self.a_up_box = QDoubleSpinBox()
-        self.a_up_box.setRange(0, 1e10)
-        self.a_up_box.setDecimals(6)
-        self.a_up_box.setValue(0.05)
-        params_layout.addRow("A_UPP:", self.a_up_box)
-
-        # F_1 parameter
-        self.f_1_box = QDoubleSpinBox()
-        self.f_1_box.setRange(0, 1e10)
-        self.f_1_box.setDecimals(6)
-        self.f_1_box.setValue(100.0)
-        params_layout.addRow("F_1:", self.f_1_box)
-
-        # F_2 parameter
-        self.f_2_box = QDoubleSpinBox()
-        self.f_2_box.setRange(0, 1e10)
-        self.f_2_box.setDecimals(6)
-        self.f_2_box.setValue(100.0)
-        params_layout.addRow("F_2:", self.f_2_box)
-
-        # OMEGA_DC parameter
-        self.omega_dc_box = QDoubleSpinBox()
-        self.omega_dc_box.setRange(0, 1e10)
-        self.omega_dc_box.setDecimals(6)
-        self.omega_dc_box.setValue(5000.0)
-        params_layout.addRow("Ω_DC:", self.omega_dc_box)
-
-        # ZETA_DC parameter
-        self.zeta_dc_box = QDoubleSpinBox()
-        self.zeta_dc_box.setRange(0, 1e10)
-        self.zeta_dc_box.setDecimals(6)
-        self.zeta_dc_box.setValue(0.01)
-        params_layout.addRow("ζ_DC:", self.zeta_dc_box)
-        
-        main_layout.addWidget(params_group)
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-
-    def create_dva_parameters_tab(self):
-        """Create the DVA parameters tab"""
-        self.dva_tab = QWidget()
-        layout = QVBoxLayout(self.dva_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Add button to apply optimized DVA parameters from GA
-        apply_optimized_container = QWidget()
-        apply_optimized_layout = QHBoxLayout(apply_optimized_container)
-        apply_optimized_layout.setContentsMargins(0, 0, 0, 10)
-        
-        apply_optimized_button = QPushButton("Apply Optimized DVA Parameters")
-        apply_optimized_button.setToolTip("Apply the best parameters from the last optimization run")
-        apply_optimized_button.clicked.connect(self.apply_optimized_dva_parameters)
-        apply_optimized_layout.addWidget(apply_optimized_button)
-        
-        apply_optimizer_combo = QComboBox()
-        apply_optimizer_combo.addItems(["Genetic Algorithm (GA)", "Particle Swarm (PSO)", 
-                                       "Differential Evolution (DE)", "Simulated Annealing (SA)", 
-                                       "CMA-ES", "Reinforcement Learning (RL)"])
-        apply_optimized_layout.addWidget(apply_optimizer_combo)
-        self.dva_optimizer_combo = apply_optimizer_combo
-        
-        main_layout.addWidget(apply_optimized_container)
-        
-        # BETA parameters group
-        beta_group = QGroupBox("β (beta) Parameters")
-        beta_form = QFormLayout(beta_group)
-        self.beta_boxes = []
-        for i in range(15):
-            b = QDoubleSpinBox()
-            b.setRange(-1e6, 1e6)
-            b.setDecimals(6)
-            self.beta_boxes.append(b)
-            beta_form.addRow(f"β_{i+1}:", b)
-        main_layout.addWidget(beta_group)
-
-        # LAMBDA parameters group
-        lambda_group = QGroupBox("λ (lambda) Parameters")
-        lambda_form = QFormLayout(lambda_group)
-        self.lambda_boxes = []
-        for i in range(15):
-            l = QDoubleSpinBox()
-            l.setRange(-1e6, 1e6)
-            l.setDecimals(6)
-            self.lambda_boxes.append(l)
-            lambda_form.addRow(f"λ_{i+1}:", l)
-        main_layout.addWidget(lambda_group)
-
-        # MU parameters group
-        mu_group = QGroupBox("μ (mu) Parameters")
-        mu_form = QFormLayout(mu_group)
-        self.mu_dva_boxes = []
-        for i in range(3):
-            m = QDoubleSpinBox()
-            m.setRange(-1e6, 1e6)
-            m.setDecimals(6)
-            self.mu_dva_boxes.append(m)
-            mu_form.addRow(f"μ_{i+1}:", m)
-        main_layout.addWidget(mu_group)
-
-        # NU parameters group
-        nu_group = QGroupBox("ν (nu) Parameters")
-        nu_form = QFormLayout(nu_group)
-        self.nu_dva_boxes = []
-        for i in range(15):
-            n = QDoubleSpinBox()
-            n.setRange(-1e6, 1e6)
-            n.setDecimals(6)
-            self.nu_dva_boxes.append(n)
-            nu_form.addRow(f"ν_{i+1}:", n)
-        main_layout.addWidget(nu_group)
-        
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-    
-    def create_target_weights_tab(self):
-        """Create the targets and weights tab"""
-        self.tw_tab = QWidget()
-        layout = QVBoxLayout(self.tw_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Initialize dictionaries to store all target and weight spinboxes
-        self.mass_target_spins = {}
-        self.mass_weight_spins = {}
-
-        # Create groups for each mass
-        for mass_num in range(1, 6):
-            mass_group = QGroupBox(f"Mass {mass_num} Targets & Weights")
-            mg_layout = QVBoxLayout(mass_group)
-
-            # Peak values group
-            peak_group = QGroupBox("Peak Values & Weights")
-            peak_form = QFormLayout(peak_group)
-            for peak_num in range(1, 5):
-                pv = QDoubleSpinBox()
-                pv.setRange(0, 1e6)
-                pv.setDecimals(6)
-                wv = QDoubleSpinBox()
-                wv.setRange(0, 1e3)
-                wv.setDecimals(6)
-                
-                peak_form.addRow(f"Peak Value {peak_num}:", pv)
-                peak_form.addRow(f"Weight Peak Value {peak_num}:", wv)
-                
-                self.mass_target_spins[f"peak_value_{peak_num}_m{mass_num}"] = pv
-                self.mass_weight_spins[f"peak_value_{peak_num}_m{mass_num}"] = wv
-            mg_layout.addWidget(peak_group)
-            
-            # Peak positions group (in a separate section)
-            peak_pos_group = QGroupBox("Peak Positions & Weights")
-            peak_pos_form = QFormLayout(peak_pos_group)
-            for peak_num in range(1, 6):  # Note: 1-5 (not 1-4)
-                pp = QDoubleSpinBox()
-                pp.setRange(0, 1e6)
-                pp.setDecimals(6)
-                wpp = QDoubleSpinBox()
-                wpp.setRange(0, 1e3)
-                wpp.setDecimals(6)
-                
-                peak_pos_form.addRow(f"Peak Position {peak_num}:", pp)
-                peak_pos_form.addRow(f"Weight Peak Position {peak_num}:", wpp)
-                
-                self.mass_target_spins[f"peak_position_{peak_num}_m{mass_num}"] = pp
-                self.mass_weight_spins[f"peak_position_{peak_num}_m{mass_num}"] = wpp
-            mg_layout.addWidget(peak_pos_group)
-
-            # Bandwidth group
-            bw_group = QGroupBox("Bandwidth Targets & Weights")
-            bw_form = QFormLayout(bw_group)
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    bw = QDoubleSpinBox()
-                    bw.setRange(0, 1e6)
-                    bw.setDecimals(6)
-                    wbw = QDoubleSpinBox()
-                    wbw.setRange(0, 1e3)
-                    wbw.setDecimals(6)
-                    bw_form.addRow(f"Bandwidth {i}-{j}:", bw)
-                    bw_form.addRow(f"Weight Bandwidth {i}-{j}:", wbw)
-                    self.mass_target_spins[f"bandwidth_{i}_{j}_m{mass_num}"] = bw
-                    self.mass_weight_spins[f"bandwidth_{i}_{j}_m{mass_num}"] = wbw
-            mg_layout.addWidget(bw_group)
-
-            # Slope group
-            slope_group = QGroupBox("Slope Targets & Weights")
-            slope_form = QFormLayout(slope_group)
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    s = QDoubleSpinBox()
-                    s.setRange(-1e6, 1e6)
-                    s.setDecimals(6)
-                    ws = QDoubleSpinBox()
-                    ws.setRange(0, 1e3)
-                    ws.setDecimals(6)
-                    slope_form.addRow(f"Slope {i}-{j}:", s)
-                    slope_form.addRow(f"Weight Slope {i}-{j}:", ws)
-                    self.mass_target_spins[f"slope_{i}_{j}_m{mass_num}"] = s
-                    self.mass_weight_spins[f"slope_{i}_{j}_m{mass_num}"] = ws
-            mg_layout.addWidget(slope_group)
-
-            # Area under curve group
-            auc_group = QGroupBox("Area Under Curve & Weight")
-            auc_form = QFormLayout(auc_group)
-            auc = QDoubleSpinBox()
-            auc.setRange(0, 1e6)
-            auc.setDecimals(6)
-            wauc = QDoubleSpinBox()
-            wauc.setRange(0, 1e3)
-            wauc.setDecimals(6)
-            auc_form.addRow("Area Under Curve:", auc)
-            auc_form.addRow("Weight Area Under Curve:", wauc)
-            self.mass_target_spins[f"area_under_curve_m{mass_num}"] = auc
-            self.mass_weight_spins[f"area_under_curve_m{mass_num}"] = wauc
-            mg_layout.addWidget(auc_group)
-
-            mg_layout.addStretch()
-            main_layout.addWidget(mass_group)
-        
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-    
-    def create_omega_sensitivity_tab(self):
-        """Create the Omega points sensitivity analysis tab"""
-        self.omega_sensitivity_tab = QWidget()
-        layout = QVBoxLayout(self.omega_sensitivity_tab)
-
-        # Create tabs for parameters and visualization
-        self.sensitivity_tabs = ModernQTabWidget()
-        layout.addWidget(self.sensitivity_tabs)
-        
-        # --------- PARAMETERS TAB ---------
-        params_tab = QWidget()
-        params_layout = QVBoxLayout(params_tab)
-        
-        # Create a scroll area for parameters
-        params_scroll_area = QScrollArea()
-        params_scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget for parameters
-        params_container = QWidget()
-        params_main_layout = QVBoxLayout(params_container)
-        
-        # Introduction group
-        intro_group = QGroupBox("Omega Points Sensitivity Analysis")
-        intro_layout = QVBoxLayout(intro_group)
-        
-        info_text = QLabel(
-            "This tool analyzes how the number of frequency points affects slope calculations in "
-            "the Frequency Response Function (FRF). It helps identify the minimum number of points "
-            "needed for stable results by incrementally increasing the frequency resolution and "
-            "observing the convergence of slope values."
-        )
-        info_text.setWordWrap(True)
-        intro_layout.addWidget(info_text)
-        
-        # Parameters group
-        params_group = QGroupBox("Analysis Parameters")
-        params_group_layout = QFormLayout(params_group)
-        
-        # Initial points
-        self.sensitivity_initial_points = QSpinBox()
-        self.sensitivity_initial_points.setRange(50, 5000)
-        self.sensitivity_initial_points.setValue(100)
-        params_group_layout.addRow("Initial Ω Points:", self.sensitivity_initial_points)
-        
-        # Maximum points
-        self.sensitivity_max_points = QSpinBox()
-        self.sensitivity_max_points.setRange(500, 1000000000)  # Allow very large values (10^9)
-        self.sensitivity_max_points.setValue(2000)
-        params_group_layout.addRow("Maximum Ω Points:", self.sensitivity_max_points)
-        
-        # Step size
-        self.sensitivity_step_size = QSpinBox()
-        self.sensitivity_step_size.setRange(10, 100000)  # Allow larger step sizes
-        self.sensitivity_step_size.setValue(1000)  # Increased default for large ranges
-        params_group_layout.addRow("Step Size:", self.sensitivity_step_size)
-        
-        # Convergence threshold
-        self.sensitivity_threshold = QDoubleSpinBox()
-        self.sensitivity_threshold.setRange(1e-10, 0.1)
-        self.sensitivity_threshold.setDecimals(10)
-        self.sensitivity_threshold.setSingleStep(1e-10)
-        self.sensitivity_threshold.setValue(0.01)
-        params_group_layout.addRow("Convergence Threshold:", self.sensitivity_threshold)
-        
-        # Max iterations
-        self.sensitivity_max_iterations = QSpinBox()
-        self.sensitivity_max_iterations.setRange(5, 1000000)  # Allow extremely high iteration counts
-        self.sensitivity_max_iterations.setValue(200)  # Set default to 200 to support larger ranges
-        params_group_layout.addRow("Maximum Iterations:", self.sensitivity_max_iterations)
-        
-        # Mass of interest
-        self.sensitivity_mass = QComboBox()
-        for i in range(1, 6):
-            self.sensitivity_mass.addItem(f"mass_{i}")
-        params_group_layout.addRow("Mass of Interest:", self.sensitivity_mass)
-        
-        # Plot results checkbox
-        self.sensitivity_plot_results = QCheckBox("Generate Convergence Plots")
-        self.sensitivity_plot_results.setChecked(True)
-        params_group_layout.addRow(self.sensitivity_plot_results)
-        
-        # Use optimal points checkbox
-        self.sensitivity_use_optimal = QCheckBox("Use Optimal Points in FRF Analysis")
-        self.sensitivity_use_optimal.setChecked(True)
-        params_group_layout.addRow(self.sensitivity_use_optimal)
-        
-        # Results group
-        self.sensitivity_results_group = QGroupBox("Analysis Results")
-        self.sensitivity_results_layout = QVBoxLayout(self.sensitivity_results_group)
-        
-        self.sensitivity_results_text = QTextEdit()
-        self.sensitivity_results_text.setReadOnly(True)
-        self.sensitivity_results_layout.addWidget(self.sensitivity_results_text)
-        
-        # Run button container
-        run_container = QWidget()
-        run_layout = QHBoxLayout(run_container)
-        run_layout.setContentsMargins(0, 20, 0, 0)  # Add some top margin
-        
-        # Add stretch to push button to center
-        run_layout.addStretch()
-        
-        # Create and style the Run Analysis button
-        self.run_sensitivity_btn = QPushButton("Run Sensitivity Analysis")
-        self.run_sensitivity_btn.setObjectName("primary-button")
-        self.run_sensitivity_btn.setMinimumWidth(200)
-        self.run_sensitivity_btn.setMinimumHeight(40)
-        self.run_sensitivity_btn.clicked.connect(self.run_omega_sensitivity)
-        self.run_sensitivity_btn.setStyleSheet("""
-            QPushButton#primary-button {
-                background-color: #4B67F0;
-                color: white;
-                border: none;
-                border-radius: 4px;
-                padding: 8px 20px;
-                font-weight: bold;
-                font-size: 14px;
-            }
-            QPushButton#primary-button:hover {
-                background-color: #3B57E0;
-            }
-            QPushButton#primary-button:pressed {
-                background-color: #2B47D0;
-            }
-        """)
-        run_layout.addWidget(self.run_sensitivity_btn)
-        
-        # Add stretch to push button to center
-        run_layout.addStretch()
-        
-        # Add all groups to main layout
-        params_main_layout.addWidget(intro_group)
-        params_main_layout.addWidget(params_group)
-        params_main_layout.addWidget(self.sensitivity_results_group)
-        params_main_layout.addWidget(run_container)
-        params_main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        params_scroll_area.setWidget(params_container)
-        params_layout.addWidget(params_scroll_area)
-        
-        # --------- VISUALIZATION TABS ---------
-        # Create visualization tabs widget
-        self.vis_tabs = ModernQTabWidget()
-        
-        # Common control panel for both visualization tabs
-        vis_control_panel = QWidget()
-        vis_control_layout = QHBoxLayout(vis_control_panel)
-        
-        # Save plot button
-        self.sensitivity_save_plot_btn = QPushButton("Save Current Plot")
-        self.sensitivity_save_plot_btn.setEnabled(False)
-        self.sensitivity_save_plot_btn.clicked.connect(self.save_sensitivity_plot)
-        vis_control_layout.addWidget(self.sensitivity_save_plot_btn)
-        
-        # Refresh plot button
-        self.sensitivity_refresh_plot_btn = QPushButton("Refresh Plots")
-        self.sensitivity_refresh_plot_btn.setEnabled(False)
-        self.sensitivity_refresh_plot_btn.clicked.connect(self.refresh_sensitivity_plot)
-        vis_control_layout.addWidget(self.sensitivity_refresh_plot_btn)
-        
-        # --------- CONVERGENCE PLOT TAB ---------
-        convergence_tab = QWidget()
-        convergence_layout = QVBoxLayout(convergence_tab)
-        
-        # Add control panel to layout
-        convergence_layout.addWidget(vis_control_panel)
-        
-        # Create figure canvas for convergence plot
-        self.convergence_fig = Figure(figsize=(10, 6))
-        self.convergence_canvas = FigureCanvas(self.convergence_fig)
-        self.convergence_canvas.setMinimumHeight(450)
-        self.convergence_toolbar = NavigationToolbar(self.convergence_canvas, convergence_tab)
-        
-        # Add canvas and toolbar to layout
-        convergence_layout.addWidget(self.convergence_canvas)
-        convergence_layout.addWidget(self.convergence_toolbar)
-        
-        # --------- RELATIVE CHANGE PLOT TAB ---------
-        rel_change_tab = QWidget()
-        rel_change_layout = QVBoxLayout(rel_change_tab)
-        
-        # Create figure canvas for relative change plot
-        self.rel_change_fig = Figure(figsize=(10, 6))
-        self.rel_change_canvas = FigureCanvas(self.rel_change_fig)
-        self.rel_change_canvas.setMinimumHeight(450)
-        self.rel_change_toolbar = NavigationToolbar(self.rel_change_canvas, rel_change_tab)
-        
-        # Add canvas and toolbar to layout
-        rel_change_layout.addWidget(self.rel_change_canvas)
-        rel_change_layout.addWidget(self.rel_change_toolbar)
-        
-        # No data message (added to both tabs)
-        self.convergence_no_data_label = QLabel("Run the sensitivity analysis to generate visualization")
-        self.convergence_no_data_label.setAlignment(Qt.AlignCenter)
-        self.convergence_no_data_label.setStyleSheet("color: #888; font-style: italic; font-size: 14px;")
-        convergence_layout.addWidget(self.convergence_no_data_label)
-        
-        self.rel_change_no_data_label = QLabel("Run the sensitivity analysis to generate visualization")
-        self.rel_change_no_data_label.setAlignment(Qt.AlignCenter)
-        self.rel_change_no_data_label.setStyleSheet("color: #888; font-style: italic; font-size: 14px;")
-        rel_change_layout.addWidget(self.rel_change_no_data_label)
-        
-        # Add tabs to the visualization tabs widget
-        self.vis_tabs.addTab(convergence_tab, "Slope Convergence")
-        self.vis_tabs.addTab(rel_change_tab, "Relative Change")
-        
-        # Create main visualization container tab
-        vis_tab = QWidget()
-        vis_layout = QVBoxLayout(vis_tab)
-        vis_layout.addWidget(self.vis_tabs)
-        self.sobol_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: Sobol Analysis Settings --------------------
-        sobol_settings_tab = QWidget()
-        sobol_settings_layout = QFormLayout(sobol_settings_tab)
-
-        self.num_samples_line = QLineEdit()
-        self.num_samples_line.setPlaceholderText("e.g. 32,64,128")
-        sobol_settings_layout.addRow("Num Samples List:", self.num_samples_line)
-
-        self.n_jobs_spin = QSpinBox()
-        self.n_jobs_spin.setRange(1, 64)
-        self.n_jobs_spin.setValue(5)
-        sobol_settings_layout.addRow("Number of Jobs (n_jobs):", self.n_jobs_spin)
-
-        # Add a small Run Sobol button in the settings sub-tab
-        self.hyper_run_sobol_button = QPushButton("Run Sobol")
-        self.hyper_run_sobol_button.setFixedWidth(100)
-        self.hyper_run_sobol_button.clicked.connect(self._run_sobol_implementation)
-        sobol_settings_layout.addRow("Run Sobol:", self.hyper_run_sobol_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        dva_param_tab = QWidget()
-        dva_param_layout = QVBoxLayout(dva_param_tab)
-
-        self.dva_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.dva_param_table.setRowCount(len(dva_parameters))
-        self.dva_param_table.setColumnCount(5)
-        self.dva_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.dva_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.dva_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.dva_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_fixed(state, r))
-            self.dva_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(-1e6, 1e6)
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setEnabled(False)
-            self.dva_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(-1e6, 1e6)
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setEnabled(True)
-            self.dva_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(-1e6, 1e6)
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setEnabled(True)
-            self.dva_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-            # Default ranges
-            if param.startswith("beta_") or param.startswith("lambda_") or param.startswith("nu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(2.5)
-            elif param.startswith("mu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(0.75)
-            else:
-                lower_bound_spin.setValue(0.0)
-                upper_bound_spin.setValue(1.0)
-
-        dva_param_layout.addWidget(self.dva_param_table)
-
-        # -------------------- Sub-tab 3: Sobol Results --------------------
-        sobol_results_tab = QWidget()
-        sobol_results_layout = QVBoxLayout(sobol_results_tab)
-        
-        # Results text area
-        self.sobol_results_text = QTextEdit()
-        self.sobol_results_text.setReadOnly(True)
-        self.sobol_results_text.setStyleSheet("font-family: monospace;")
-        sobol_results_layout.addWidget(self.sobol_results_text)
-        
-        # Placeholder for Sobol plots
-        plot_container = QWidget()
-        plot_layout = QVBoxLayout(plot_container)
-        
-        # Plotting controls section
-        plot_controls = QWidget()
-        plot_controls_layout = QHBoxLayout(plot_controls)
-        
-        self.sobol_combo = QComboBox()
-        self.sobol_combo.currentIndexChanged.connect(self.update_sobol_plot)
-        plot_controls_layout.addWidget(QLabel("Plot Type:"))
-        plot_controls_layout.addWidget(self.sobol_combo)
-        
-        save_button = QPushButton("Save Results")
-        save_button.clicked.connect(self.save_sobol_results)
-        plot_controls_layout.addWidget(save_button)
-        
-        plot_layout.addWidget(plot_controls)
-        
-        # Figure canvas
-        self.sobol_figure = Figure(figsize=(4, 4))
-        self.sobol_canvas = FigureCanvas(self.sobol_figure)
-        self.sobol_canvas.setMinimumHeight(300)
-        toolbar = NavigationToolbar(self.sobol_canvas, self)
-        
-        plot_layout.addWidget(toolbar)
-        plot_layout.addWidget(self.sobol_canvas)
-        
-        sobol_results_layout.addWidget(plot_container)
-        
-        # Initialize plots dictionary
-        self.sobol_plots = {}
-
-        # Add sub-tabs to the Sobol tab widget
-        self.sobol_sub_tabs.addTab(sobol_settings_tab, "Analysis Settings")
-        self.sobol_sub_tabs.addTab(dva_param_tab, "DVA Parameters")
-        self.sobol_sub_tabs.addTab(sobol_results_tab, "Results")
-
-        # Add the Sobol sub-tabs widget to the main Sobol tab layout
-        layout.addWidget(self.sobol_sub_tabs)
-        self.sobol_tab.setLayout(layout)
-
 
     def _run_sobol_implementation(self):
         """Run Sobol sensitivity analysis - main implementation"""
@@ -1708,1909 +1064,7 @@ class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin
             self.zeta_dc_box.value()
         )
         
-    def get_dva_params(self):
-        """Get the DVA parameters in a dictionary format"""
-        dva_params = {
-            "beta": [box.value() for box in self.beta_boxes],
-            "lambda": [box.value() for box in self.lambda_boxes],
-            "mu": [box.value() for box in self.mu_dva_boxes],
-            "nu": [box.value() for box in self.nu_dva_boxes]
-        }
-        return dva_params
         
-    def get_target_values_weights(self):
-        """
-        Get the target values and weights for all masses.
-        
-        Returns:
-            tuple: (target_values_dict, weights_dict) containing targets and weights for all masses
-        """
-        target_values_dict = {}
-        weights_dict = {}
-
-        for mass_num in range(1, 6):
-            t_dict = {}
-            w_dict = {}
-            # Handle peak values (1-4)
-            for peak_num in range(1, 5):
-                pv_key = f"peak_value_{peak_num}_m{mass_num}"
-                if pv_key in self.mass_target_spins:
-                    t_dict[f"peak_value_{peak_num}"] = self.mass_target_spins[pv_key].value()
-                    w_dict[f"peak_value_{peak_num}"] = self.mass_weight_spins[pv_key].value()
-            
-            # Handle peak positions (1-5)
-            for peak_num in range(1, 6):
-                pp_key = f"peak_position_{peak_num}_m{mass_num}"
-                if pp_key in self.mass_target_spins:
-                    t_dict[f"peak_position_{peak_num}"] = self.mass_target_spins[pp_key].value()
-                    w_dict[f"peak_position_{peak_num}"] = self.mass_weight_spins[pp_key].value()
-
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    bw_key = f"bandwidth_{i}_{j}_m{mass_num}"
-                    if bw_key in self.mass_target_spins:
-                        t_dict[f"bandwidth_{i}_{j}"] = self.mass_target_spins[bw_key].value()
-                        w_dict[f"bandwidth_{i}_{j}"] = self.mass_weight_spins[bw_key].value()
-
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    slope_key = f"slope_{i}_{j}_m{mass_num}"
-                    if slope_key in self.mass_target_spins:
-                        t_dict[f"slope_{i}_{j}"] = self.mass_target_spins[slope_key].value()
-                        w_dict[f"slope_{i}_{j}"] = self.mass_weight_spins[slope_key].value()
-
-            auc_key = f"area_under_curve_m{mass_num}"
-            if auc_key in self.mass_target_spins:
-                t_dict["area_under_curve"] = self.mass_target_spins[auc_key].value()
-                w_dict["area_under_curve"] = self.mass_weight_spins[auc_key].value()
-
-            target_values_dict[f"mass_{mass_num}"] = t_dict
-            weights_dict[f"mass_{mass_num}"] = w_dict
-
-        return target_values_dict, weights_dict
-        
-    def create_de_tab(self):
-        """Create the differential evolution optimization tab"""
-        self.de_tab = QWidget()
-        layout = QVBoxLayout(self.de_tab)
-        
-        # Create sub-tabs widget
-        self.de_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: DE Hyperparameters --------------------
-        de_hyper_tab = QWidget()
-        de_hyper_layout = QFormLayout(de_hyper_tab)
-
-        self.de_pop_size_box = QSpinBox()
-        self.de_pop_size_box.setRange(1, 10000)
-        self.de_pop_size_box.setValue(50)
-
-        self.de_num_generations_box = QSpinBox()
-        self.de_num_generations_box.setRange(1, 10000)
-        self.de_num_generations_box.setValue(100)
-
-        self.de_F_box = QDoubleSpinBox()
-        self.de_F_box.setRange(0, 2)
-        self.de_F_box.setValue(0.8)
-        self.de_F_box.setDecimals(3)
-
-        self.de_CR_box = QDoubleSpinBox()
-        self.de_CR_box.setRange(0, 1)
-        self.de_CR_box.setValue(0.7)
-        self.de_CR_box.setDecimals(3)
-
-        self.de_tol_box = QDoubleSpinBox()
-        self.de_tol_box.setRange(0, 1e6)
-        self.de_tol_box.setValue(1e-3)
-        self.de_tol_box.setDecimals(6)
-
-        self.de_alpha_box = QDoubleSpinBox()
-        self.de_alpha_box.setRange(0.0, 10.0)
-        self.de_alpha_box.setDecimals(4)
-        self.de_alpha_box.setSingleStep(0.01)
-        self.de_alpha_box.setValue(0.01)
-        
-        # New smoothness penalty parameter
-        self.de_beta_box = QDoubleSpinBox()
-        self.de_beta_box.setRange(0.0, 10.0)
-        self.de_beta_box.setDecimals(4)
-        self.de_beta_box.setSingleStep(0.01)
-        self.de_beta_box.setValue(0.0)
-        self.de_beta_box.setToolTip("Parameter for smoothness penalty (0 = no smoothness penalty)")
-
-        de_hyper_layout.addRow("Population Size:", self.de_pop_size_box)
-        de_hyper_layout.addRow("Number of Generations:", self.de_num_generations_box)
-        de_hyper_layout.addRow("Mutation Factor (F):", self.de_F_box)
-        de_hyper_layout.addRow("Crossover Rate (CR):", self.de_CR_box)
-        de_hyper_layout.addRow("Tolerance (tol):", self.de_tol_box)
-        de_hyper_layout.addRow("Sparsity Penalty (alpha):", self.de_alpha_box)
-        de_hyper_layout.addRow("Smoothness Penalty (beta):", self.de_beta_box)
-
-        # Add a small Run DE button in the hyperparameters sub-tab
-        self.hyper_run_de_button = QPushButton("Run DE")
-        self.hyper_run_de_button.setFixedWidth(100)
-        self.hyper_run_de_button.clicked.connect(self.run_de)
-        de_hyper_layout.addRow("Run DE:", self.hyper_run_de_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        de_param_tab = QWidget()
-        de_param_layout = QVBoxLayout(de_param_tab)
-
-        self.de_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.de_param_table.setRowCount(len(dva_parameters))
-        self.de_param_table.setColumnCount(5)
-        self.de_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.de_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.de_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.de_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_de_fixed(state, r))
-            self.de_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(-1e6, 1e6)
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setEnabled(False)
-            self.de_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(-1e6, 1e6)
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setEnabled(True)
-            self.de_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(-1e6, 1e6)
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setEnabled(True)
-            self.de_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-            # Default ranges
-            if param.startswith("beta_") or param.startswith("lambda_") or param.startswith("nu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(2.5)
-            elif param.startswith("mu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(0.75)
-            else:
-                lower_bound_spin.setValue(0.0)
-                upper_bound_spin.setValue(1.0)
-
-        de_param_layout.addWidget(self.de_param_table)
-
-        # -------------------- Sub-tab 3: Results --------------------
-        de_results_tab = QWidget()
-        de_results_layout = QVBoxLayout(de_results_tab)
-        
-        self.de_results_text = QTextEdit()
-        self.de_results_text.setReadOnly(True)
-        de_results_layout.addWidget(QLabel("DE Optimization Results:"))
-        de_results_layout.addWidget(self.de_results_text)
-
-        # -------------------- Sub-tab 4: Advanced Settings --------------------
-        de_advanced_tab = QWidget()
-        de_advanced_layout = QVBoxLayout(de_advanced_tab)
-        
-        # Create scrollable area for advanced settings
-        advanced_scroll = QScrollArea()
-        advanced_scroll.setWidgetResizable(True)
-        advanced_scroll_content = QWidget()
-        advanced_scroll_layout = QVBoxLayout(advanced_scroll_content)
-        advanced_scroll.setWidget(advanced_scroll_content)
-        
-        # DE strategy selection
-        strategy_group = QGroupBox("DE Strategy")
-        strategy_layout = QVBoxLayout(strategy_group)
-        
-        self.de_strategy_combo = QComboBox()
-        self.de_strategy_combo.addItems([
-            "rand/1 (Standard DE)", 
-            "rand/2", 
-            "best/1", 
-            "best/2",
-            "current-to-best/1", 
-            "current-to-rand/1"
-        ])
-        self.de_strategy_combo.setToolTip("Different mutation strategies for creating donor vectors")
-        strategy_layout.addWidget(self.de_strategy_combo)
-        
-        advanced_scroll_layout.addWidget(strategy_group)
-        
-        # Adaptive methods
-        adaptive_group = QGroupBox("Adaptive Parameter Control")
-        adaptive_layout = QFormLayout(adaptive_group)
-        
-        self.de_adaptive_method_combo = QComboBox()
-        self.de_adaptive_method_combo.addItems([
-            "none (Fixed Parameters)",
-            "jitter (Small Random Variation)",
-            "dither (Random F per Generation)",
-            "sade (Self-adaptive DE)",
-            "jade (Adaptive DE with Archive)",
-            "success-history (Success-based Adaptation)"
-        ])
-        self.de_adaptive_method_combo.setToolTip("Methods for automatically adapting control parameters during optimization")
-        adaptive_layout.addRow("Adaptation Method:", self.de_adaptive_method_combo)
-        
-        # JADE parameters
-        jade_frame = QGroupBox("JADE Parameters")
-        jade_layout = QFormLayout(jade_frame)
-        
-        self.de_jade_c_box = QDoubleSpinBox()
-        self.de_jade_c_box.setRange(0.01, 1.0)
-        self.de_jade_c_box.setValue(0.1)
-        self.de_jade_c_box.setDecimals(2)
-        jade_layout.addRow("Adaptation Rate (c):", self.de_jade_c_box)
-        
-        adaptive_layout.addRow("", jade_frame)
-        
-        # SaDE parameters
-        sade_frame = QGroupBox("SaDE Parameters")
-        sade_layout = QFormLayout(sade_frame)
-        
-        self.de_sade_lp_box = QSpinBox()
-        self.de_sade_lp_box.setRange(1, 1000)
-        self.de_sade_lp_box.setValue(50)
-        sade_layout.addRow("Learning Period:", self.de_sade_lp_box)
-        
-        self.de_sade_memory_box = QSpinBox()
-        self.de_sade_memory_box.setRange(1, 100)
-        self.de_sade_memory_box.setValue(20)
-        sade_layout.addRow("Memory Size:", self.de_sade_memory_box)
-        
-        adaptive_layout.addRow("", sade_frame)
-        
-        # Dither parameters
-        dither_frame = QGroupBox("Dither Parameters")
-        dither_layout = QFormLayout(dither_frame)
-        
-        self.de_dither_f_min_box = QDoubleSpinBox()
-        self.de_dither_f_min_box.setRange(0.1, 0.9)
-        self.de_dither_f_min_box.setValue(0.4)
-        self.de_dither_f_min_box.setDecimals(2)
-        dither_layout.addRow("F Minimum:", self.de_dither_f_min_box)
-        
-        self.de_dither_f_max_box = QDoubleSpinBox()
-        self.de_dither_f_max_box.setRange(0.1, 2.0)
-        self.de_dither_f_max_box.setValue(0.9)
-        self.de_dither_f_max_box.setDecimals(2)
-        dither_layout.addRow("F Maximum:", self.de_dither_f_max_box)
-        
-        adaptive_layout.addRow("", dither_frame)
-        
-        advanced_scroll_layout.addWidget(adaptive_group)
-        
-        # Constraint handling
-        constraint_group = QGroupBox("Constraint Handling")
-        constraint_layout = QVBoxLayout(constraint_group)
-        
-        self.de_constraint_handling_combo = QComboBox()
-        self.de_constraint_handling_combo.addItems([
-            "penalty (Apply Penalty)",
-            "reflection (Reflect at Bounds)",
-            "projection (Project to Bounds)"
-        ])
-        self.de_constraint_handling_combo.setToolTip("Method for handling parameter constraints")
-        constraint_layout.addWidget(self.de_constraint_handling_combo)
-        
-        advanced_scroll_layout.addWidget(constraint_group)
-        
-        # Termination criteria
-        termination_group = QGroupBox("Termination Criteria")
-        termination_layout = QFormLayout(termination_group)
-        
-        self.de_stagnation_box = QSpinBox()
-        self.de_stagnation_box.setRange(10, 10000)
-        self.de_stagnation_box.setValue(100)
-        termination_layout.addRow("Max Generations without Improvement:", self.de_stagnation_box)
-        
-        self.de_min_diversity_box = QDoubleSpinBox()
-        self.de_min_diversity_box.setRange(1e-10, 1.0)
-        self.de_min_diversity_box.setValue(1e-6)
-        self.de_min_diversity_box.setDecimals(10)
-        termination_layout.addRow("Minimum Population Diversity:", self.de_min_diversity_box)
-        
-        advanced_scroll_layout.addWidget(termination_group)
-        
-        # Diversity preservation
-        diversity_group = QGroupBox("Diversity Preservation")
-        diversity_layout = QFormLayout(diversity_group)
-        
-        self.de_diversity_checkbox = QCheckBox()
-        self.de_diversity_checkbox.setChecked(False)
-        diversity_layout.addRow("Enable Diversity Preservation:", self.de_diversity_checkbox)
-        
-        self.de_diversity_threshold_box = QDoubleSpinBox()
-        self.de_diversity_threshold_box.setRange(1e-6, 1.0)
-        self.de_diversity_threshold_box.setValue(0.01)
-        self.de_diversity_threshold_box.setDecimals(4)
-        diversity_layout.addRow("Diversity Threshold:", self.de_diversity_threshold_box)
-        
-        advanced_scroll_layout.addWidget(diversity_group)
-        
-        # Add multiple run settings
-        multi_run_group = QGroupBox("Multiple Runs")
-        multi_run_layout = QFormLayout(multi_run_group)
-        
-        self.de_num_runs_box = QSpinBox()
-        self.de_num_runs_box.setRange(1, 100)
-        self.de_num_runs_box.setValue(1)
-        self.de_num_runs_box.setToolTip("Number of independent optimization runs to perform")
-        multi_run_layout.addRow("Number of Runs:", self.de_num_runs_box)
-        
-        advanced_scroll_layout.addWidget(multi_run_group)
-        
-        # Add multi-run progress bar
-        self.de_multi_run_progress_bar = QProgressBar()
-        self.de_multi_run_progress_bar.setFormat("Run %v/%m")
-        self.de_multi_run_progress_bar.hide()
-        advanced_scroll_layout.addWidget(self.de_multi_run_progress_bar)
-        
-        # Parallel processing
-        parallel_group = QGroupBox("Parallel Processing")
-        parallel_layout = QFormLayout(parallel_group)
-        
-        self.de_parallel_checkbox = QCheckBox()
-        self.de_parallel_checkbox.setChecked(False)
-        parallel_layout.addRow("Use Parallel Processing:", self.de_parallel_checkbox)
-        
-        self.de_processes_box = QSpinBox()
-        self.de_processes_box.setRange(1, 64)
-        # Use multiprocessing properly
-        import multiprocessing
-        self.de_processes_box.setValue(max(1, multiprocessing.cpu_count() - 1))
-        self.de_processes_box.setEnabled(False)
-        parallel_layout.addRow("Number of Processes:", self.de_processes_box)
-        
-        # Connect parallel checkbox to enable/disable processes box
-        self.de_parallel_checkbox.stateChanged.connect(
-            lambda state: self.de_processes_box.setEnabled(state == Qt.Checked)
-        )
-        
-        advanced_scroll_layout.addWidget(parallel_group)
-        
-        # Random seed
-        seed_group = QGroupBox("Random Seed")
-        seed_layout = QFormLayout(seed_group)
-        
-        self.de_seed_checkbox = QCheckBox()
-        self.de_seed_checkbox.setChecked(False)
-        seed_layout.addRow("Use Fixed Seed:", self.de_seed_checkbox)
-        
-        self.de_seed_box = QSpinBox()
-        self.de_seed_box.setRange(0, 1000000)
-        self.de_seed_box.setValue(42)
-        self.de_seed_box.setEnabled(False)
-        seed_layout.addRow("Random Seed:", self.de_seed_box)
-        
-        # Connect seed checkbox to enable/disable seed box
-        self.de_seed_checkbox.stateChanged.connect(
-            lambda state: self.de_seed_box.setEnabled(state == Qt.Checked)
-        )
-        
-        advanced_scroll_layout.addWidget(seed_group)
-        
-        # Hyperparameter tuning
-        tuning_group = QGroupBox("Hyperparameter Tuning")
-        tuning_layout = QVBoxLayout(tuning_group)
-        
-        self.de_tune_button = QPushButton("Tune DE Hyperparameters")
-        self.de_tune_button.setToolTip("Run automatic hyperparameter tuning to find optimal settings")
-        self.de_tune_button.clicked.connect(self.tune_de_hyperparameters)
-        tuning_layout.addWidget(self.de_tune_button)
-        
-        advanced_scroll_layout.addWidget(tuning_group)
-        
-        # Add stretch to push widgets to the top
-        advanced_scroll_layout.addStretch()
-        
-        # Add scroll area to advanced tab
-        de_advanced_layout.addWidget(advanced_scroll)
-        
-        # -------------------- Sub-tab 5: Visualization --------------------
-        de_viz_tab = QWidget()
-        de_viz_layout = QVBoxLayout(de_viz_tab)
-        
-        # Add save button
-        save_container = QWidget()
-        save_layout = QHBoxLayout(save_container)
-        save_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.de_viz_save_button = QPushButton("Save Plot")
-        self.de_viz_save_button.clicked.connect(self.save_de_visualization)
-        save_layout.addWidget(self.de_viz_save_button)
-        save_layout.addStretch()
-        
-        de_viz_layout.addWidget(save_container)
-        
-        # Create tabs for different visualizations
-        self.de_viz_tabs = QTabWidget()
-        
-        # Create tabs for different visualizations
-        de_violin_tab = QWidget()
-        de_violin_layout = QVBoxLayout(de_violin_tab)
-        self.de_violin_plot_widget = QWidget()
-        de_violin_layout.addWidget(self.de_violin_plot_widget)
-        
-        de_convergence_tab = QWidget()
-        de_convergence_layout = QVBoxLayout(de_convergence_tab)
-        self.de_convergence_plot_widget = QWidget()
-        de_convergence_layout.addWidget(self.de_convergence_plot_widget)
-        
-        de_diversity_tab = QWidget()
-        de_diversity_layout = QVBoxLayout(de_diversity_tab)
-        self.de_diversity_plot_widget = QWidget()
-        de_diversity_layout.addWidget(self.de_diversity_plot_widget)
-        
-        de_adaptation_tab = QWidget()
-        de_adaptation_layout = QVBoxLayout(de_adaptation_tab)
-        self.de_adaptation_plot_widget = QWidget()
-        de_adaptation_layout.addWidget(self.de_adaptation_plot_widget)
-        
-        de_param_evolution_tab = QWidget()
-        de_param_evolution_layout = QVBoxLayout(de_param_evolution_tab)
-        self.de_param_evolution_plot_widget = QWidget()
-        de_param_evolution_layout.addWidget(self.de_param_evolution_plot_widget)
-        
-        de_correlation_tab = QWidget()
-        de_correlation_layout = QVBoxLayout(de_correlation_tab)
-        self.de_correlation_plot_widget = QWidget()
-        de_correlation_layout.addWidget(self.de_correlation_plot_widget)
-        
-        # Add all visualization tabs
-        self.de_viz_tabs.addTab(de_violin_tab, "Multi-Run Statistics")
-        self.de_viz_tabs.addTab(de_convergence_tab, "Convergence History")
-        self.de_viz_tabs.addTab(de_diversity_tab, "Population Diversity")
-        self.de_viz_tabs.addTab(de_adaptation_tab, "Control Parameter Adaptation")
-        self.de_viz_tabs.addTab(de_param_evolution_tab, "Parameter Evolution")
-        self.de_viz_tabs.addTab(de_correlation_tab, "Parameter Correlation")
-        
-        # Add the visualization tabs to the layout
-        de_viz_layout.addWidget(self.de_viz_tabs)
-        
-        # Connect tab change to update function
-        self.de_viz_tabs.currentChanged.connect(self.update_de_visualization)
-        
-        # Add all sub-tabs to the DE tab widget
-        self.de_sub_tabs.addTab(de_hyper_tab, "DE Settings")
-        self.de_sub_tabs.addTab(de_param_tab, "DVA Parameters")
-        self.de_sub_tabs.addTab(de_results_tab, "Results")
-        self.de_sub_tabs.addTab(de_advanced_tab, "Advanced Settings")
-        self.de_sub_tabs.addTab(de_viz_tab, "Visualization")
-
-        # Add the DE sub-tabs widget to the main DE tab layout
-        layout.addWidget(self.de_sub_tabs)
-        self.de_tab.setLayout(layout)
-        
-    def toggle_de_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a DE parameter row"""
-        if table is None:
-            table = self.de_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-        
-    def run_de(self):
-        """Run the differential evolution optimization"""
-        self.status_bar.showMessage("Running DE optimization...")
-        self.results_text.append("DE optimization started...")
-        
-        try:
-            # Retrieve DE parameters from the GUI
-            pop_size = self.de_pop_size_box.value()
-            num_generations = self.de_num_generations_box.value()
-            mutation_factor = self.de_F_box.value()
-            crossover_rate = self.de_CR_box.value()
-            tol = self.de_tol_box.value()
-            alpha = self.de_alpha_box.value()
-            beta = self.de_beta_box.value()
-            
-            # Get advanced DE options
-            strategy_idx = self.de_strategy_combo.currentIndex()
-            strategy_names = ["rand/1", "rand/2", "best/1", "best/2", "current-to-best/1", "current-to-rand/1"]
-            strategy = strategy_names[strategy_idx]
-            
-            adaptive_idx = self.de_adaptive_method_combo.currentIndex()
-            adaptive_names = ["none", "jitter", "dither", "sade", "jade", "success-history"]
-            adaptive_method = adaptive_names[adaptive_idx]
-            
-            constraint_idx = self.de_constraint_handling_combo.currentIndex()
-            constraint_names = ["penalty", "reflection", "projection"]
-            constraint_handling = constraint_names[constraint_idx]
-            
-            use_parallel = self.de_parallel_checkbox.isChecked()
-            n_processes = self.de_processes_box.value() if use_parallel else None
-            
-            use_seed = self.de_seed_checkbox.isChecked()
-            seed = self.de_seed_box.value() if use_seed else None
-            
-            diversity_preservation = self.de_diversity_checkbox.isChecked()
-            
-            # Get number of runs
-            num_runs = self.de_num_runs_box.value()
-            
-            # Show multi-run progress bar if doing multiple runs
-            if num_runs > 1:
-                self.de_multi_run_progress_bar.setRange(0, num_runs)
-                self.de_multi_run_progress_bar.setValue(0)
-                self.de_multi_run_progress_bar.show()
-            else:
-                self.de_multi_run_progress_bar.hide()
-            
-            # Prepare adaptive parameters based on selected method
-            adaptive_params = {}
-            if adaptive_method == "jade":
-                adaptive_params["c"] = self.de_jade_c_box.value()
-            elif adaptive_method == "sade":
-                adaptive_params["LP"] = self.de_sade_lp_box.value()
-                adaptive_params["memory_size"] = self.de_sade_memory_box.value()
-            elif adaptive_method == "dither":
-                adaptive_params["F_min"] = self.de_dither_f_min_box.value()
-                adaptive_params["F_max"] = self.de_dither_f_max_box.value()
-            
-            # Diversity preservation parameters
-            if diversity_preservation:
-                adaptive_params["diversity_threshold"] = self.de_diversity_threshold_box.value()
-            
-            # Prepare termination criteria
-            termination_criteria = {
-                "max_generations": num_generations,
-                "tol": tol,
-                "stagnation_limit": self.de_stagnation_box.value(),
-                "min_diversity": self.de_min_diversity_box.value()
-            }
-
-            de_dva_parameters = []
-            row_count = self.de_param_table.rowCount()
-            for row in range(row_count):
-                param_name = self.de_param_table.item(row, 0).text()
-                fixed_widget = self.de_param_table.cellWidget(row, 1)
-                fixed = fixed_widget.isChecked()
-                if fixed:
-                    fixed_value_widget = self.de_param_table.cellWidget(row, 2)
-                    fv = fixed_value_widget.value()
-                    de_dva_parameters.append((param_name, fv, fv, True))
-                else:
-                    lower_bound_widget = self.de_param_table.cellWidget(row, 3)
-                    upper_bound_widget = self.de_param_table.cellWidget(row, 4)
-                    lb = lower_bound_widget.value()
-                    ub = upper_bound_widget.value()
-                    if lb > ub:
-                        QMessageBox.warning(self, "Input Error",
-                                            f"For parameter {param_name}, lower bound is greater than upper bound.")
-                        return
-                    de_dva_parameters.append((param_name, lb, ub, False))
-
-            # Get main system parameters
-            main_params = self.get_main_system_params()
-
-            # Get target values and weights
-            target_values, weights = self.get_target_values_weights()
-
-            # Get frequency range values
-            omega_start_val = self.omega_start_box.value()
-            omega_end_val = self.omega_end_box.value()
-            omega_points_val = self.omega_points_box.value()
-
-            # Create and start DEWorker with enhanced parameters
-            self.de_worker = DEWorker(
-                main_params=main_params,
-                target_values_dict=target_values,
-                weights_dict=weights,
-                omega_start=omega_start_val,
-                omega_end=omega_end_val,
-                omega_points=omega_points_val,
-                de_pop_size=pop_size,
-                de_num_generations=num_generations,
-                de_F=mutation_factor,
-                de_CR=crossover_rate,
-                de_tol=tol,
-                de_parameter_data=de_dva_parameters,
-                alpha=alpha,
-                beta=beta,
-                strategy=strategy,
-                adaptive_method=adaptive_method,
-                adaptive_params=adaptive_params,
-                termination_criteria=termination_criteria,
-                use_parallel=use_parallel,
-                n_processes=n_processes,
-                seed=seed,
-                record_statistics=True,
-                constraint_handling=constraint_handling,
-                diversity_preservation=diversity_preservation,
-                num_runs=num_runs  # Add number of runs parameter
-            )
-            
-            self.de_worker.finished.connect(self.handle_de_finished)
-            self.de_worker.error.connect(self.handle_de_error)
-            self.de_worker.update.connect(self.handle_de_update)
-            self.de_worker.progress.connect(self.handle_de_progress)
-            self.de_worker.multi_run_progress.connect(self.handle_de_multi_run_progress)  # Connect multi-run progress signal
-            
-            # Disable both run DE buttons to prevent multiple runs
-            self.hyper_run_de_button.setEnabled(False)
-            self.run_de_button.setEnabled(False)
-            
-            self.de_results_text.clear()
-            self.de_results_text.append("Running DE optimization...")
-            
-            # Initialize progress bar if not exists
-            if not hasattr(self, 'de_progress_bar'):
-                self.de_progress_bar = QProgressBar()
-                self.de_progress_bar.setRange(0, num_generations)
-                self.de_progress_bar.setFormat("%v/%m gen - Best fitness: %p%")
-                self.de_sub_tabs.widget(3).layout().addWidget(self.de_progress_bar)
-            else:
-                self.de_progress_bar.setRange(0, num_generations)
-                self.de_progress_bar.setValue(0)
-            
-            self.de_worker.start()
-            
-        except Exception as e:
-            self.handle_de_error(str(e))
-        
-    
-    def handle_de_finished(self, results, best_individual, parameter_names, best_fitness, statistics):
-        """Handle the completion of DE optimization"""
-        # Re-enable both run DE buttons
-        self.hyper_run_de_button.setEnabled(True)
-        self.run_de_button.setEnabled(True)
-        
-        self.de_results_text.append("\nDE Completed.\n")
-        self.de_results_text.append("Best individual parameters:")
-
-        for name, val in zip(parameter_names, best_individual):
-            self.de_results_text.append(f"{name}: {val}")
-        self.de_results_text.append(f"\nBest fitness: {best_fitness:.6f}")
-
-        singular_response = results.get('singular_response', None)
-        if singular_response is not None:
-            self.de_results_text.append(f"\nSingular response of best individual: {singular_response}")
-
-        self.de_results_text.append("\nFull Results:")
-        for section, data in results.items():
-            self.de_results_text.append(f"{section}: {data}")
-            
-        # Store optimization statistics for visualization
-        self.de_statistics = statistics
-        
-        # Update visualization tab
-        self.update_de_visualization()
-            
-        self.status_bar.showMessage("DE optimization completed")
-        
-        # Store results for later use in comparative visualization
-        if "singular_response" in results and results["singular_response"] is not None:
-            # Instead of using the missing method, just store the results for potential comparison
-            if not hasattr(self, 'frf_comparison_results'):
-                self.frf_comparison_results = {}
-            self.frf_comparison_results["DE"] = results
-            
-        # Store best parameters for potential application
-        self.current_de_best_params = best_individual
-        self.current_de_parameter_names = parameter_names
-
-    def handle_de_error(self, err):
-        """Handle errors during DE optimization"""
-        # Re-enable both run DE buttons
-        self.hyper_run_de_button.setEnabled(True)
-        self.run_de_button.setEnabled(True)
-        
-        QMessageBox.warning(self, "DE Error", f"Error during DE optimization: {err}")
-        self.de_results_text.append(f"\nError running DE: {err}")
-        self.status_bar.showMessage("DE optimization failed")
-
-    def handle_de_update(self, msg):
-        """Handle progress updates from DE worker"""
-        self.de_results_text.append(msg)
-        
-    def handle_de_progress(self, generation, best_fitness, diversity):
-        """Handle progress updates from DE worker"""
-        if hasattr(self, 'de_progress_bar'):
-            self.de_progress_bar.setValue(generation)
-            self.de_progress_bar.setFormat(f"{generation}/{self.de_progress_bar.maximum()} gen - Best: {best_fitness:.6f}")
-            
-    def handle_de_multi_run_progress(self, current_run, total_runs):
-        """Handle progress updates for multiple runs"""
-        self.de_multi_run_progress_bar.setValue(current_run)
-        self.status_bar.showMessage(f"Running DE optimization - Run {current_run}/{total_runs}")
-            
-    def update_de_visualization(self):
-        """Update the DE visualization based on selected tab"""
-        if not hasattr(self, 'de_statistics'):
-            return
-            
-        # Get current tab
-        current_tab = self.de_viz_tabs.currentWidget()
-        
-        # Clear the current widget's layout
-        if current_tab.layout():
-            for i in reversed(range(current_tab.layout().count())): 
-                widget = current_tab.layout().itemAt(i).widget()
-                if widget:
-                    widget.setParent(None)
-        
-        # Create figure and canvas for the current visualization
-        fig = Figure(figsize=(8, 6))
-        canvas = FigureCanvas(fig)
-        toolbar = NavigationToolbar(canvas, current_tab)
-        
-        # Add toolbar and canvas to layout
-        current_tab.layout().addWidget(toolbar)
-        current_tab.layout().addWidget(canvas)
-        
-        # Get tab name
-        tab_name = self.de_viz_tabs.tabText(self.de_viz_tabs.currentIndex())
-        
-        if tab_name == "Multi-Run Statistics" and hasattr(self.de_statistics, 'run_best_fitnesses'):
-            # Create violin plots for multi-run statistics
-            n_params = len(self.current_de_parameter_names)
-            n_rows = (n_params + 2 + 1) // 2  # Parameters + fitness + convergence, 2 columns
-            
-            # Create subplots
-            for i, param_name in enumerate(self.current_de_parameter_names):
-                ax = fig.add_subplot(n_rows, 2, i + 1)
-                param_values = self.de_statistics.parameter_distributions[param_name]
-                sns.violinplot(data=param_values, ax=ax)
-                ax.set_title(f'Distribution of {param_name}')
-                ax.set_ylabel('Parameter Value')
-            
-            # Add fitness distribution plot
-            ax = fig.add_subplot(n_rows, 2, n_params + 1)
-            sns.violinplot(data=self.de_statistics.run_best_fitnesses, ax=ax)
-            ax.set_title('Distribution of Best Fitness Values')
-            ax.set_ylabel('Fitness Value')
-            
-            # Add convergence generation distribution plot
-            ax = fig.add_subplot(n_rows, 2, n_params + 2)
-            sns.violinplot(data=self.de_statistics.run_convergence_gens, ax=ax)
-            ax.set_title('Distribution of Convergence Generations')
-            ax.set_ylabel('Generation')
-            
-        elif tab_name == "Convergence History":
-            ax = fig.add_subplot(111)
-            generations = self.de_statistics.generations
-            best_fitness = self.de_statistics.best_fitness_history
-            mean_fitness = self.de_statistics.mean_fitness_history
-            
-            ax.plot(generations, best_fitness, 'b-', label='Best Fitness')
-            ax.plot(generations, mean_fitness, 'r--', label='Mean Fitness')
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Fitness Value')
-            ax.set_title('Convergence History')
-            ax.legend()
-            ax.grid(True)
-            
-        elif tab_name == "Population Diversity":
-            ax = fig.add_subplot(111)
-            generations = self.de_statistics.generations
-            diversity = self.de_statistics.diversity_history
-            
-            ax.plot(generations, diversity, 'g-')
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Population Diversity')
-            ax.set_title('Population Diversity Over Time')
-            ax.grid(True)
-            
-        elif tab_name == "Control Parameter Adaptation":
-            if hasattr(self.de_statistics, 'f_values') and hasattr(self.de_statistics, 'cr_values'):
-                ax1 = fig.add_subplot(211)
-                ax2 = fig.add_subplot(212)
-                
-                generations = self.de_statistics.generations
-                f_values = self.de_statistics.f_values
-                cr_values = self.de_statistics.cr_values
-                
-                ax1.plot(generations, f_values, 'b-')
-                ax1.set_xlabel('Generation')
-                ax1.set_ylabel('F Value')
-                ax1.set_title('Mutation Factor (F) Adaptation')
-                ax1.grid(True)
-                
-                ax2.plot(generations, cr_values, 'r-')
-                ax2.set_xlabel('Generation')
-                ax2.set_ylabel('CR Value')
-                ax2.set_title('Crossover Rate (CR) Adaptation')
-                ax2.grid(True)
-            
-        elif tab_name == "Parameter Evolution":
-            ax = fig.add_subplot(111)
-            generations = self.de_statistics.generations
-            param_means = np.array(self.de_statistics.parameter_mean_history)
-            param_stds = np.array(self.de_statistics.parameter_std_history)
-            
-            for i, param_name in enumerate(self.current_de_parameter_names):
-                mean = param_means[:, i]
-                std = param_stds[:, i]
-                ax.plot(generations, mean, label=param_name)
-                ax.fill_between(generations, mean - std, mean + std, alpha=0.2)
-            
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Parameter Value')
-            ax.set_title('Parameter Evolution Over Time')
-            ax.legend(bbox_to_anchor=(1.05, 1), loc='upper left')
-            ax.grid(True)
-            
-        elif tab_name == "Parameter Correlation":
-            if len(self.current_de_parameter_names) > 1:
-                param_data = {}
-                for i, param_name in enumerate(self.current_de_parameter_names):
-                    param_data[param_name] = np.array(self.de_statistics.parameter_mean_history)[:, i]
-                
-                df = pd.DataFrame(param_data)
-                sns.heatmap(df.corr(), annot=True, cmap='coolwarm', ax=fig.add_subplot(111))
-                plt.title('Parameter Correlation Matrix')
-        
-        # Adjust layout and draw
-        fig.tight_layout()
-        canvas.draw()
-        if not hasattr(self, 'de_statistics') or self.de_statistics is None:
-            return
-            
-        viz_type = self.de_viz_combo.currentText()
-        
-        # Clear the figure
-        self.de_viz_fig.clear()
-        
-        if viz_type == "Convergence History":
-            ax = self.de_viz_fig.add_subplot(111)
-            ax.plot(self.de_statistics.generations, self.de_statistics.best_fitness_history, 'b-', label='Best Fitness')
-            ax.plot(self.de_statistics.generations, self.de_statistics.mean_fitness_history, 'r--', label='Mean Fitness')
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Fitness Value')
-            ax.set_title('Convergence History')
-            ax.legend()
-            ax.grid(True)
-            
-        elif viz_type == "Population Diversity":
-            ax = self.de_viz_fig.add_subplot(111)
-            ax.plot(self.de_statistics.generations, self.de_statistics.diversity_history, 'g-')
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Population Diversity')
-            ax.set_title('Population Diversity History')
-            ax.grid(True)
-            
-        elif viz_type == "Control Parameter Adaptation":
-            ax = self.de_viz_fig.add_subplot(111)
-            ax.plot(self.de_statistics.generations, self.de_statistics.f_values, 'b-', label='F')
-            ax.plot(self.de_statistics.generations, self.de_statistics.cr_values, 'r-', label='CR')
-            ax.set_xlabel('Generation')
-            ax.set_ylabel('Parameter Value')
-            ax.set_title('Control Parameter Adaptation')
-            ax.legend()
-            ax.grid(True)
-            
-        elif viz_type == "Parameter Evolution":
-            if len(self.de_statistics.parameter_mean_history) > 0:
-                param_means = np.array(self.de_statistics.parameter_mean_history)
-                param_stds = np.array(self.de_statistics.parameter_std_history)
-                
-                n_params = param_means.shape[1]
-                if n_params > 10:  # Too many parameters to show all at once
-                    # Show just a sample of parameters
-                    param_indices = np.linspace(0, n_params-1, 6, dtype=int)
-                    
-                    fig = self.de_viz_fig
-                    fig.subplots_adjust(hspace=0.4, wspace=0.4)
-                    
-                    for i, idx in enumerate(param_indices):
-                        ax = fig.add_subplot(2, 3, i+1)
-                        ax.plot(self.de_statistics.generations, param_means[:, idx], 'b-')
-                        ax.fill_between(
-                            self.de_statistics.generations,
-                            param_means[:, idx] - param_stds[:, idx],
-                            param_means[:, idx] + param_stds[:, idx],
-                            alpha=0.2
-                        )
-                        param_name = self.current_de_parameter_names[idx] if hasattr(self, 'current_de_parameter_names') else f"Param {idx}"
-                        ax.set_title(f"{param_name}")
-                        ax.grid(True)
-                    
-                    fig.suptitle("Parameter Evolution (Sample)", fontsize=12)
-                else:
-                    # Show all parameters
-                    ax = self.de_viz_fig.add_subplot(111)
-                    for i in range(n_params):
-                        ax.plot(self.de_statistics.generations, param_means[:, i], label=f"Param {i}")
-                    ax.set_xlabel('Generation')
-                    ax.set_ylabel('Parameter Value')
-                    ax.set_title('Parameter Evolution')
-                    if n_params <= 5:  # Only show legend if not too many parameters
-                        ax.legend()
-                    ax.grid(True)
-            
-        elif viz_type == "Parameter Correlation":
-            if len(self.de_statistics.parameter_mean_history) > 0:
-                param_means = np.array(self.de_statistics.parameter_mean_history)
-                
-                # Create correlation matrix
-                corr_matrix = np.corrcoef(param_means.T)
-                
-                # Create heatmap
-                ax = self.de_viz_fig.add_subplot(111)
-                cax = ax.matshow(corr_matrix, cmap='coolwarm', vmin=-1, vmax=1)
-                self.de_viz_fig.colorbar(cax)
-                
-                # Set axis labels with parameter names if available
-                if hasattr(self, 'current_de_parameter_names'):
-                    # Use shorter parameter names for readability
-                    short_names = [name.split('_')[-1] for name in self.current_de_parameter_names]
-                    if len(short_names) > 15:
-                        # Too many parameters, show indices instead
-                        ax.set_title("Parameter Correlation Matrix")
-                    else:
-                        ax.set_xticks(np.arange(len(short_names)))
-                        ax.set_yticks(np.arange(len(short_names)))
-                        ax.set_xticklabels(short_names, rotation=90)
-                        ax.set_yticklabels(short_names)
-                        ax.set_title("Parameter Correlation Matrix")
-        
-        # Refresh the canvas
-        self.de_viz_canvas.draw()
-        
-    def save_de_visualization(self):
-        """Save the current DE visualization plot"""
-        if not hasattr(self, 'de_statistics') or self.de_statistics is None:
-            QMessageBox.warning(self, "No Data", "No visualization data available to save.")
-            return
-            
-        viz_type = self.de_viz_combo.currentText()
-        file_path, _ = QFileDialog.getSaveFileName(
-            self, "Save Visualization", "", "PNG Files (*.png);;PDF Files (*.pdf);;All Files (*)"
-        )
-        
-        if file_path:
-            self.de_viz_fig.savefig(file_path, dpi=300, bbox_inches='tight')
-            self.status_bar.showMessage(f"Visualization saved to {file_path}")
-            
-    def tune_de_hyperparameters(self):
-        """Run hyperparameter tuning for DE"""
-        reply = QMessageBox.question(
-            self, 'Hyperparameter Tuning',
-            'Hyperparameter tuning will evaluate multiple combinations of DE parameters '
-            'which may take a significant amount of time. Continue?',
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No
-        )
-        
-        if reply == QMessageBox.No:
-            return
-            
-        try:
-            # Get necessary parameters for tuning
-            main_params = self.get_main_system_params()
-            target_values, weights = self.get_target_values_weights()
-            omega_start_val = self.omega_start_box.value()
-            omega_end_val = self.omega_end_box.value()
-            omega_points_val = self.omega_points_box.value()
-            
-            # Get DVA parameters
-            de_dva_parameters = []
-            row_count = self.de_param_table.rowCount()
-            for row in range(row_count):
-                param_name = self.de_param_table.item(row, 0).text()
-                fixed_widget = self.de_param_table.cellWidget(row, 1)
-                fixed = fixed_widget.isChecked()
-                if fixed:
-                    fixed_value_widget = self.de_param_table.cellWidget(row, 2)
-                    fv = fixed_value_widget.value()
-                    de_dva_parameters.append((param_name, fv, fv, True))
-                else:
-                    lower_bound_widget = self.de_param_table.cellWidget(row, 3)
-                    upper_bound_widget = self.de_param_table.cellWidget(row, 4)
-                    lb = lower_bound_widget.value()
-                    ub = upper_bound_widget.value()
-                    de_dva_parameters.append((param_name, lb, ub, False))
-            
-            # Create a progress dialog
-            progress_dialog = QDialog(self)
-            progress_dialog.setWindowTitle("DE Hyperparameter Tuning")
-            progress_dialog.setMinimumWidth(400)
-            
-            dialog_layout = QVBoxLayout(progress_dialog)
-            dialog_layout.addWidget(QLabel("Running hyperparameter tuning. This may take some time..."))
-            
-            tuning_progress_text = QTextEdit()
-            tuning_progress_text.setReadOnly(True)
-            dialog_layout.addWidget(tuning_progress_text)
-            
-            progress_bar = QProgressBar()
-            progress_bar.setRange(0, 0)  # Indeterminate progress
-            dialog_layout.addWidget(progress_bar)
-            
-            cancel_button = QPushButton("Cancel")
-            cancel_button.clicked.connect(progress_dialog.reject)
-            dialog_layout.addWidget(cancel_button)
-            
-            # Create a worker thread for tuning
-            class TuningThread(QThread):
-                finished_signal = pyqtSignal(dict)
-                update_signal = pyqtSignal(str)
-                
-                def __init__(self, main_params, target_values, weights, omega_start, 
-                             omega_end, omega_points, de_parameter_data):
-                    super().__init__()
-                    self.main_params = main_params
-                    self.target_values = target_values
-                    self.weights = weights
-                    self.omega_start = omega_start
-                    self.omega_end = omega_end
-                    self.omega_points = omega_points
-                    self.de_parameter_data = de_parameter_data
-                    
-                def run(self):
-                    try:
-                        # Redirect print output to emit as signal
-                        original_print = print
-                        def print_override(*args, **kwargs):
-                            message = " ".join(map(str, args))
-                            self.update_signal.emit(message)
-                            original_print(*args, **kwargs)
-                        
-                        __builtins__['print'] = print_override
-                        
-                        # Run tuning with fewer trials for UI responsiveness
-                        best_params = DEWorker.tune_hyperparameters(
-                            main_params=self.main_params,
-                            target_values_dict=self.target_values,
-                            weights_dict=self.weights,
-                            omega_start=self.omega_start,
-                            omega_end=self.omega_end,
-                            omega_points=self.omega_points,
-                            de_parameter_data=self.de_parameter_data,
-                            n_trials=5,  # Reduced for UI
-                            parallel=True
-                        )
-                        
-                        # Restore original print
-                        __builtins__['print'] = original_print
-                        
-                        self.finished_signal.emit(best_params)
-                    except Exception as e:
-                        self.update_signal.emit(f"Error in tuning: {str(e)}")
-                        self.finished_signal.emit({})
-            
-            # Create and start the tuning thread
-            tuning_thread = TuningThread(
-                main_params, target_values, weights, 
-                omega_start_val, omega_end_val, omega_points_val,
-                de_dva_parameters
-            )
-            
-            # Connect signals
-            tuning_thread.update_signal.connect(lambda msg: tuning_progress_text.append(msg))
-            tuning_thread.finished_signal.connect(lambda result: self._apply_tuning_results(result, progress_dialog))
-            
-            # Start the thread and show dialog
-            tuning_thread.start()
-            progress_dialog.exec_()
-            
-            # Ensure thread is terminated if dialog is closed
-            if tuning_thread.isRunning():
-                tuning_thread.terminate()
-                tuning_thread.wait()
-                
-        except Exception as e:
-            QMessageBox.warning(self, "Tuning Error", f"Error during hyperparameter tuning: {str(e)}")
-    
-    def _apply_tuning_results(self, best_params, dialog):
-        """Apply the results of hyperparameter tuning to the UI"""
-        if not best_params:
-            QMessageBox.warning(self, "Tuning Error", "Hyperparameter tuning failed or was cancelled.")
-            dialog.accept()
-            return
-            
-        # Update UI controls with best parameters
-        self.de_pop_size_box.setValue(best_params.get("pop_size", 50))
-        self.de_F_box.setValue(best_params.get("F", 0.5))
-        self.de_CR_box.setValue(best_params.get("CR", 0.7))
-        
-        # Update strategy combo box
-        if "strategy" in best_params:
-            strategy_map = {
-                "rand/1": 0,
-                "rand/2": 1,
-                "best/1": 2,
-                "best/2": 3,
-                "current-to-best/1": 4,
-                "current-to-rand/1": 5
-            }
-            strategy_value = best_params["strategy"].value if hasattr(best_params["strategy"], "value") else best_params["strategy"]
-            if strategy_value in strategy_map:
-                self.de_strategy_combo.setCurrentIndex(strategy_map[strategy_value])
-        
-        # Show summary of results
-        QMessageBox.information(
-            self, "Tuning Complete",
-            f"Hyperparameter tuning completed.\n\n"
-            f"Best parameters:\n"
-            f"Population Size: {best_params.get('pop_size', 'N/A')}\n"
-            f"Mutation Factor (F): {best_params.get('F', 'N/A')}\n"
-            f"Crossover Rate (CR): {best_params.get('CR', 'N/A')}\n"
-            f"Strategy: {best_params.get('strategy', 'N/A')}\n\n"
-            f"Average Fitness: {best_params.get('avg_fitness', 'N/A')}\n"
-            f"Average Convergence Generation: {best_params.get('avg_convergence_gen', 'N/A')}"
-        )
-        
-        dialog.accept()
-        
-    def create_sa_tab(self):
-        """Create the simulated annealing optimization tab"""
-        self.sa_tab = QWidget()
-        layout = QVBoxLayout(self.sa_tab)
-        
-        # Create sub-tabs widget
-        self.sa_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: SA Hyperparameters --------------------
-        sa_hyper_tab = QWidget()
-        sa_hyper_layout = QFormLayout(sa_hyper_tab)
-
-        self.sa_initial_temp_box = QDoubleSpinBox()
-        self.sa_initial_temp_box.setRange(0, 1e6)
-        self.sa_initial_temp_box.setValue(1000)
-        self.sa_initial_temp_box.setDecimals(2)
-
-        self.sa_cooling_rate_box = QDoubleSpinBox()
-        self.sa_cooling_rate_box.setRange(0, 1)
-        self.sa_cooling_rate_box.setValue(0.95)
-        self.sa_cooling_rate_box.setDecimals(3)
-
-        self.sa_num_iterations_box = QSpinBox()
-        self.sa_num_iterations_box.setRange(1, 10000)
-        self.sa_num_iterations_box.setValue(1000)
-
-        self.sa_tol_box = QDoubleSpinBox()
-        self.sa_tol_box.setRange(0, 1e6)
-        self.sa_tol_box.setValue(1e-3)
-        self.sa_tol_box.setDecimals(6)
-
-        self.sa_alpha_box = QDoubleSpinBox()
-        self.sa_alpha_box.setRange(0.0, 10.0)
-        self.sa_alpha_box.setDecimals(4)
-        self.sa_alpha_box.setSingleStep(0.01)
-        self.sa_alpha_box.setValue(0.01)
-
-        sa_hyper_layout.addRow("Initial Temperature:", self.sa_initial_temp_box)
-        sa_hyper_layout.addRow("Cooling Rate:", self.sa_cooling_rate_box)
-        sa_hyper_layout.addRow("Number of Iterations:", self.sa_num_iterations_box)
-        sa_hyper_layout.addRow("Tolerance (tol):", self.sa_tol_box)
-        sa_hyper_layout.addRow("Sparsity Penalty (alpha):", self.sa_alpha_box)
-
-        # Add a small Run SA button in the hyperparameters sub-tab
-        self.hyper_run_sa_button = QPushButton("Run SA")
-        self.hyper_run_sa_button.setFixedWidth(100)
-        self.hyper_run_sa_button.clicked.connect(self.run_sa)
-        sa_hyper_layout.addRow("Run SA:", self.hyper_run_sa_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        sa_param_tab = QWidget()
-        sa_param_layout = QVBoxLayout(sa_param_tab)
-
-        self.sa_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.sa_param_table.setRowCount(len(dva_parameters))
-        self.sa_param_table.setColumnCount(5)
-        self.sa_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.sa_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.sa_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        # Set up table rows
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.sa_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_sa_fixed(state, r))
-            self.sa_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(-1e6, 1e6)
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setEnabled(False)
-            self.sa_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(-1e6, 1e6)
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setEnabled(True)
-            self.sa_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(-1e6, 1e6)
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setEnabled(True)
-            self.sa_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-            # Default ranges
-            if param.startswith("beta_") or param.startswith("lambda_") or param.startswith("nu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(2.5)
-            elif param.startswith("mu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(0.75)
-            else:
-                lower_bound_spin.setValue(0.0)
-                upper_bound_spin.setValue(1.0)
-
-        sa_param_layout.addWidget(self.sa_param_table)
-
-        # -------------------- Sub-tab 3: Results --------------------
-        sa_results_tab = QWidget()
-        sa_results_layout = QVBoxLayout(sa_results_tab)
-        
-        self.sa_results_text = QTextEdit()
-        self.sa_results_text.setReadOnly(True)
-        sa_results_layout.addWidget(QLabel("SA Optimization Results:"))
-        sa_results_layout.addWidget(self.sa_results_text)
-
-        # Add all sub-tabs to the SA tab widget
-        self.sa_sub_tabs.addTab(sa_hyper_tab, "SA Settings")
-        self.sa_sub_tabs.addTab(sa_param_tab, "DVA Parameters")
-        self.sa_sub_tabs.addTab(sa_results_tab, "Results")
-
-        # Add the SA sub-tabs widget to the main SA tab layout
-        layout.addWidget(self.sa_sub_tabs)
-        self.sa_tab.setLayout(layout)
-        
-    def toggle_sa_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a SA parameter row"""
-        if table is None:
-            table = self.sa_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-        
-    def run_sa(self):
-        """Run the simulated annealing optimization"""
-        # Implementation already exists at line 2591
-        pass
-        
-    def run_cmaes(self):
-        """Run the CMA-ES optimization"""
-        # Implementation already exists at line 2840
-        pass
-        
-    def create_cmaes_tab(self):
-        """Create the CMA-ES optimization tab"""
-        self.cmaes_tab = QWidget()
-        layout = QVBoxLayout(self.cmaes_tab)
-
-        # Create sub-tabs widget
-        self.cmaes_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: CMA-ES Hyperparameters --------------------
-        cmaes_hyper_tab = QWidget()
-        cmaes_hyper_layout = QFormLayout(cmaes_hyper_tab)
-
-        self.cmaes_sigma_box = QDoubleSpinBox()
-        self.cmaes_sigma_box.setRange(0, 1e6)
-        self.cmaes_sigma_box.setValue(0.5)
-        self.cmaes_sigma_box.setDecimals(2)
-
-        self.cmaes_max_iter_box = QSpinBox()
-        self.cmaes_max_iter_box.setRange(1, 10000)
-        self.cmaes_max_iter_box.setValue(500)
-
-        self.cmaes_tol_box = QDoubleSpinBox()
-        self.cmaes_tol_box.setRange(0, 1e6)
-        self.cmaes_tol_box.setValue(1e-3)
-        self.cmaes_tol_box.setDecimals(6)
-
-        self.cmaes_alpha_box = QDoubleSpinBox()
-        self.cmaes_alpha_box.setRange(0.0, 10.0)
-        self.cmaes_alpha_box.setDecimals(4)
-        self.cmaes_alpha_box.setSingleStep(0.01)
-        self.cmaes_alpha_box.setValue(0.01)
-
-        cmaes_hyper_layout.addRow("Initial Sigma:", self.cmaes_sigma_box)
-        cmaes_hyper_layout.addRow("Max Iterations:", self.cmaes_max_iter_box)
-        cmaes_hyper_layout.addRow("Tolerance (tol):", self.cmaes_tol_box)
-        cmaes_hyper_layout.addRow("Sparsity Penalty (alpha):", self.cmaes_alpha_box)
-
-        # Add a small Run CMA-ES button in the hyperparameters sub-tab
-        self.hyper_run_cmaes_button = QPushButton("Run CMA-ES")
-        self.hyper_run_cmaes_button.setFixedWidth(100)
-        self.hyper_run_cmaes_button.clicked.connect(self.run_cmaes)
-        cmaes_hyper_layout.addRow("Run CMA-ES:", self.hyper_run_cmaes_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        cmaes_param_tab = QWidget()
-        cmaes_param_layout = QVBoxLayout(cmaes_param_tab)
-
-        self.cmaes_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.cmaes_param_table.setRowCount(len(dva_parameters))
-        self.cmaes_param_table.setColumnCount(5)
-        self.cmaes_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.cmaes_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.cmaes_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        # Set up table rows
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.cmaes_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_cmaes_fixed(state, r))
-            self.cmaes_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(-1e6, 1e6)
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setEnabled(False)
-            self.cmaes_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(-1e6, 1e6)
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setEnabled(True)
-            self.cmaes_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(-1e6, 1e6)
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setEnabled(True)
-            self.cmaes_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-            # Set default ranges based on parameter name
-            if param.startswith("beta_") or param.startswith("lambda_") or param.startswith("nu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(2.5)
-            elif param.startswith("mu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(0.75)
-            else:
-                lower_bound_spin.setValue(0.0)
-                upper_bound_spin.setValue(1.0)
-
-        cmaes_param_layout.addWidget(self.cmaes_param_table)
-
-        # -------------------- Sub-tab 3: Results --------------------
-        cmaes_results_tab = QWidget()
-        cmaes_results_layout = QVBoxLayout(cmaes_results_tab)
-        
-        self.cmaes_results_text = QTextEdit()
-        self.cmaes_results_text.setReadOnly(True)
-        cmaes_results_layout.addWidget(QLabel("CMA-ES Optimization Results:"))
-        cmaes_results_layout.addWidget(self.cmaes_results_text)
-
-        # Add all sub-tabs to the CMA-ES tab widget
-        self.cmaes_sub_tabs.addTab(cmaes_hyper_tab, "CMA-ES Settings")
-        self.cmaes_sub_tabs.addTab(cmaes_param_tab, "DVA Parameters")
-        self.cmaes_sub_tabs.addTab(cmaes_results_tab, "Results")
-
-        # Add the CMA-ES sub-tabs widget to the main CMA-ES tab layout
-        layout.addWidget(self.cmaes_sub_tabs)
-        self.cmaes_tab.setLayout(layout)
-
-    def toggle_cmaes_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a CMA-ES parameter row"""
-        if table is None:
-            table = self.cmaes_param_table
-        
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-
-    def run_cmaes(self):
-        """Run the CMA-ES optimization"""
-        # Implementation already exists at line 2840
-        pass
-
-    def handle_cmaes_finished(self, results, best_candidate, parameter_names, best_fitness):
-        """Handle the completion of CMA-ES optimization"""
-        # Re-enable both run CMA-ES buttons
-        self.hyper_run_cmaes_button.setEnabled(True)
-        self.run_cmaes_button.setEnabled(True)
-        
-        self.cmaes_results_text.append("\nCMA-ES Completed.\n")
-        self.cmaes_results_text.append("Best candidate parameters:")
-
-        for name, val in zip(parameter_names, best_candidate):
-            self.cmaes_results_text.append(f"{name}: {val}")
-        self.cmaes_results_text.append(f"\nBest fitness: {best_fitness:.6f}")
-
-        singular_response = results.get('singular_response', None)
-        if singular_response is not None:
-            self.cmaes_results_text.append(f"\nSingular response of best candidate: {singular_response}")
-
-        self.cmaes_results_text.append("\nFull Results:")
-        for section, data in results.items():
-            self.cmaes_results_text.append(f"{section}: {data}")
-            
-        self.status_bar.showMessage("CMA-ES optimization completed")
-
-    def handle_cmaes_error(self, err):
-        """Handle errors during CMA-ES optimization"""
-        # Re-enable both run CMA-ES buttons
-        self.hyper_run_cmaes_button.setEnabled(True)
-        self.run_cmaes_button.setEnabled(True)
-        
-        QMessageBox.warning(self, "CMA-ES Error", f"Error during CMA-ES optimization: {err}")
-        self.cmaes_results_text.append(f"\nError running CMA-ES: {err}")
-        self.status_bar.showMessage("CMA-ES optimization failed")
-
-    def handle_cmaes_update(self, msg):
-        """Handle progress updates from CMA-ES worker"""
-        self.cmaes_results_text.append(msg)
-
-    # Comprehensive Analysis create_rl_tab method has been removed
-
-    # RL-related method removed
-
-    # RL-related method removed
-
-    # RL-related method removed
-
-    # RL-related method removed
-    
-    def run_omega_sensitivity(self):
-        """Run the Omega points sensitivity analysis"""
-        # Get main system parameters
-        main_params = self.get_main_system_params()
-        
-        # Get DVA parameters - ensure we have 48 values (15 betas, 15 lambdas, 3 mus, 15 nus)
-        dva_params = []
-        
-        # Add beta parameters (15)
-        for i in range(15):
-            if i < len(self.beta_boxes):
-                dva_params.append(self.beta_boxes[i].value())
-            else:
-                dva_params.append(0.0)
-                
-        # Add lambda parameters (15)
-        for i in range(15):
-            if i < len(self.lambda_boxes):
-                dva_params.append(self.lambda_boxes[i].value())
-            else:
-                dva_params.append(0.0)
-                
-        # Add mu parameters (3)
-        for i in range(3):
-            if i < len(self.mu_dva_boxes):
-                dva_params.append(self.mu_dva_boxes[i].value())
-            else:
-                dva_params.append(0.0)
-                
-        # Add nu parameters (15)
-        for i in range(15):
-            if i < len(self.nu_dva_boxes):
-                dva_params.append(self.nu_dva_boxes[i].value())
-            else:
-                dva_params.append(0.0)
-        
-        # Get the omega range from the frequency tab
-        omega_start = self.omega_start_box.value()
-        omega_end = self.omega_end_box.value()
-        
-        # Check if start is less than end
-        if omega_start >= omega_end:
-            QMessageBox.warning(self, "Input Error", "Ω Start must be less than Ω End.")
-            return
-        
-        # Get sensitivity analysis parameters
-        initial_points = self.sensitivity_initial_points.value()
-        max_points = self.sensitivity_max_points.value()
-        step_size = self.sensitivity_step_size.value()
-        convergence_threshold = self.sensitivity_threshold.value()
-        max_iterations = self.sensitivity_max_iterations.value()
-        mass_of_interest = self.sensitivity_mass.currentText()
-        plot_results = self.sensitivity_plot_results.isChecked()
-        
-        # Update UI
-        self.sensitivity_results_text.clear()
-        self.sensitivity_results_text.append("Running Omega points sensitivity analysis...\n")
-        self.status_bar.showMessage("Running Omega points sensitivity analysis...")
-        
-        # Disable run button during analysis
-        self.run_sensitivity_btn.setEnabled(False)
-        
-        # Create worker for background processing
-        class SensitivityWorker(QThread):
-            finished = pyqtSignal(dict)
-            error = pyqtSignal(str)
-            
-            def __init__(self, main_params, dva_params, omega_start, omega_end, 
-                         initial_points, max_points, step_size, convergence_threshold,
-                         max_iterations, mass_of_interest, plot_results):
-                super().__init__()
-                self.main_params = main_params
-                self.dva_params = dva_params
-                self.omega_start = omega_start
-                self.omega_end = omega_end
-                self.initial_points = initial_points
-                self.max_points = max_points
-                self.step_size = step_size
-                self.convergence_threshold = convergence_threshold
-                self.max_iterations = max_iterations
-                self.mass_of_interest = mass_of_interest
-                self.plot_results = plot_results
-            
-            def run(self):
-                try:
-                    # Import the function from FRF module
-                    from modules.FRF import perform_omega_points_sensitivity_analysis
-                    
-                    # Run sensitivity analysis
-                    results = perform_omega_points_sensitivity_analysis(
-                        main_system_parameters=self.main_params,
-                        dva_parameters=self.dva_params,
-                        omega_start=self.omega_start,
-                        omega_end=self.omega_end,
-                        initial_points=self.initial_points,
-                        max_points=self.max_points,
-                        step_size=self.step_size,
-                        convergence_threshold=self.convergence_threshold,
-                        max_iterations=self.max_iterations,
-                        mass_of_interest=self.mass_of_interest,
-                        plot_results=self.plot_results
-                    )
-                    
-                    self.finished.emit(results)
-                except Exception as e:
-                    import traceback
-                    self.error.emit(f"Error in sensitivity analysis: {str(e)}\n{traceback.format_exc()}")
-
-        # Create and start the worker
-        self.sensitivity_worker = SensitivityWorker(
-            main_params, dva_params, omega_start, omega_end, 
-            initial_points, max_points, step_size, convergence_threshold,
-            max_iterations, mass_of_interest, plot_results
-        )
-        
-        # Connect signals
-        self.sensitivity_worker.finished.connect(self.handle_sensitivity_finished)
-        self.sensitivity_worker.error.connect(self.handle_sensitivity_error)
-        
-        # Start worker
-        self.sensitivity_worker.start()
-    
-    def handle_sensitivity_finished(self, results):
-        """Handle the completion of the Omega points sensitivity analysis"""
-        # Re-enable run button
-        self.run_sensitivity_btn.setEnabled(True)
-        
-        # Update status
-        self.status_bar.showMessage("Omega points sensitivity analysis completed")
-        
-        # Store results for later use in plotting
-        self.sensitivity_results = results
-        
-        # Display results
-        self.sensitivity_results_text.append("\n--- Analysis Results ---\n")
-        
-        # Show analysis outcome with detailed information
-        optimal_points = results["optimal_points"]
-        converged = results["converged"]
-        convergence_point = results.get("convergence_point")
-        all_analyzed = results.get("all_points_analyzed", False)
-        requested_max = results.get("requested_max_points", optimal_points)
-        highest_analyzed = results.get("highest_analyzed_point", optimal_points)
-        hit_iter_limit = results.get("iteration_limit_reached", False)
-        
-        # No automatic step size adjustment as per user request
-        
-        # Did the analysis reach the requested maximum points?
-        if requested_max > highest_analyzed:
-            # No, it stopped early
-            self.sensitivity_results_text.append(f"⚠️ WARNING: Analysis stopped at {highest_analyzed} points (requested maximum: {requested_max})\n")
-            
-            if hit_iter_limit:
-                self.sensitivity_results_text.append(f"   Reason: Maximum number of iterations reached ({self.sensitivity_max_iterations.value()})\n")
-                self.sensitivity_results_text.append(f"   Solution: Increase 'Maximum Iterations' parameter to analyze more points\n")
-            else:
-                self.sensitivity_results_text.append(f"   Possible reasons: calculation constraints or memory limits\n")
-                self.sensitivity_results_text.append(f"   Try using an even larger step size for higher point values\n")
-        
-        # Show convergence status
-        if converged:
-            if convergence_point == optimal_points:
-                # Converged right at the last point
-                self.sensitivity_results_text.append(f"✅ Analysis converged at {convergence_point} omega points\n")
-            else:
-                # Converged earlier but continued as requested
-                self.sensitivity_results_text.append(f"✅ Analysis converged at {convergence_point} omega points, continued to {highest_analyzed}\n")
-                
-            # Report explicitly about whether we made it to max_points
-            if all_analyzed:
-                self.sensitivity_results_text.append(f"   Successfully analyzed all requested points up to maximum: {requested_max}\n")
-        else:
-            # Did not converge anywhere
-            self.sensitivity_results_text.append(f"⚠️ Analysis did not converge at any point up to {highest_analyzed} omega points\n")
-        
-        # Show result details in a formatted table
-        self.sensitivity_results_text.append("--- Detailed Results ---")
-        self.sensitivity_results_text.append("Points | Max Slope | Relative Change")
-        self.sensitivity_results_text.append("-------|-----------|----------------")
-        
-        for i in range(len(results["omega_points"])):
-            points = results["omega_points"][i]
-            slope = results["max_slopes"][i]
-            change = results["relative_changes"][i] if i < len(results["relative_changes"]) else float('nan')
-            
-            if not np.isnan(change):
-                change_str = f"{change:.6f}"
-            else:
-                change_str = "N/A"
-                
-            self.sensitivity_results_text.append(f"{points:6d} | {slope:10.6f} | {change_str}")
-                
-        # If user selected to use optimal points, update the FRF omega points setting
-        if self.sensitivity_use_optimal.isChecked():
-            # Use the highest points value we calculated, or the requested max if we reached it
-            points_to_use = requested_max if all_analyzed else highest_analyzed
-            
-            # Update UI
-            self.omega_points_box.setValue(points_to_use)
-            self.sensitivity_results_text.append(f"\nAutomatically updated Frequency tab's Ω Points to {points_to_use}")
-            
-        # Create visualization using our improved dual-plot system
-        self.refresh_sensitivity_plot()
-            
-        # Enable the buttons for plot interaction
-        self.sensitivity_save_plot_btn.setEnabled(True)
-        self.sensitivity_refresh_plot_btn.setEnabled(True)
-
-    def handle_sensitivity_error(self, error_msg):
-        """Handle errors in the Omega points sensitivity analysis"""
-        # Re-enable run button
-        self.run_sensitivity_btn.setEnabled(True)
-        
-        # Update status
-        self.status_bar.showMessage("Omega points sensitivity analysis failed")
-        
-        # Display error message
-        self.sensitivity_results_text.append(f"\n❌ ERROR: {error_msg}")
-        
-        # Also show a message box
-        QMessageBox.critical(self, "Sensitivity Analysis Error", error_msg)
-        
-    def save_sensitivity_plot(self):
-        """Save the current sensitivity analysis plot to a file"""
-        # Determine which tab is active and save that plot
-        current_tab_idx = self.vis_tabs.currentIndex()
-        
-        if current_tab_idx == 0:  # Convergence plot
-            if not hasattr(self, 'convergence_fig') or self.convergence_fig is None:
-                QMessageBox.warning(self, "Error", "No convergence plot to save.")
-                return
-                
-            self.save_plot(self.convergence_fig, "Slope_Convergence_Analysis")
-            
-        elif current_tab_idx == 1:  # Relative change plot
-            if not hasattr(self, 'rel_change_fig') or self.rel_change_fig is None:
-                QMessageBox.warning(self, "Error", "No relative change plot to save.")
-                return
-                
-            self.save_plot(self.rel_change_fig, "Relative_Change_Analysis")
-    
-    def refresh_sensitivity_plot(self):
-        """Refresh the sensitivity analysis plots"""
-        if not hasattr(self, 'sensitivity_results') or not self.sensitivity_results:
-            QMessageBox.warning(self, "Error", "No sensitivity analysis results available to refresh.")
-            return
-            
-        # Recreate plots with stored results
-        results = self.sensitivity_results
-        
-        # ---------- CONVERGENCE PLOT ----------
-        # Clear the convergence figure
-        self.convergence_fig.clear()
-        
-        # Get the convergence data
-        omega_points = results["omega_points"]
-        max_slopes = results["max_slopes"]
-        optimal_points = results["optimal_points"]
-        
-        # Create the plot
-        ax = self.convergence_fig.add_subplot(111)
-        
-        # Add data points with connecting lines
-        ax.plot(omega_points, max_slopes, 'o-', linewidth=2.5, markersize=8, 
-                color='#4B67F0', alpha=0.8, label=f'Maximum Slope (Latest: {max_slopes[-1]:.6f})')
-        
-        # Add optimal point vertical line
-        ax.axvline(x=optimal_points, color='red', linestyle='--', linewidth=2, 
-                  label=f'Optimal: {optimal_points} points')
-        
-        # Highlight convergence point and optimal point differently
-        convergence_point = results.get("convergence_point")
-        
-        # If convergence was detected, mark it specially
-        if convergence_point is not None and convergence_point in omega_points:
-            try:
-                convergence_idx = list(omega_points).index(convergence_point)
-                convergence_y = max_slopes[convergence_idx]
-                ax.scatter(convergence_point, convergence_y, s=200, color='green', zorder=6, 
-                          marker='*', label=f'Convergence Point ({convergence_point} pts, slope: {convergence_y:.6f})')
-            except (ValueError, IndexError):
-                # Convergence point not in the list, skip highlighting
-                pass
-                
-        # Always highlight the optimal point
-        try:
-            optimal_idx = list(omega_points).index(optimal_points)
-            optimal_y = max_slopes[optimal_idx]
-            
-            # If optimal is same as convergence, don't add duplicate label
-            if optimal_points == convergence_point:
-                ax.scatter(optimal_points, optimal_y, s=250, color='purple', zorder=7, 
-                          marker='*', label=f'Optimal & Convergence Point ({optimal_points} pts)')
-            else:
-                ax.scatter(optimal_points, optimal_y, s=180, color='red', zorder=5, 
-                          marker='*', label=f'Optimal Point ({optimal_points} pts)')
-        except (ValueError, IndexError):
-            # Optimal point not in the list, skip highlighting
-            pass
-        
-        # Add grid and labels
-        ax.grid(True, linestyle='--', alpha=0.7)
-        ax.set_xlabel('Number of Omega Points', fontsize=14, fontweight='bold')
-        ax.set_ylabel('Maximum Slope Value', fontsize=14, fontweight='bold')
-        ax.set_title('Convergence of Maximum Slope with Increasing Frequency Resolution', 
-                    fontsize=16, fontweight='bold')
-                    
-        # Add a text box with key statistics
-        textstr = '\n'.join((
-            f'Summary Statistics:',
-            f'Initial Points: {omega_points[0]}',
-            f'Final Points: {omega_points[-1]}',
-            f'Optimal Points: {optimal_points}',
-            f'Initial Slope: {max_slopes[0]:.6f}',
-            f'Final Slope: {max_slopes[-1]:.6f}'
-        ))
-        props = dict(boxstyle='round,pad=0.8', facecolor='wheat', alpha=0.7)
-        ax.text(0.05, 0.05, textstr, transform=ax.transAxes, fontsize=11,
-               verticalalignment='bottom', bbox=props)
-        
-        # Create enhanced legend with better formatting
-        legend = ax.legend(fontsize=12, loc='best', frameon=True, fancybox=True, shadow=True, 
-                         title="Convergence Analysis Results", title_fontsize=14,
-                         borderpad=1.2, labelspacing=1.2, handletextpad=1.0)
-        
-        # Update the figure
-        self.convergence_fig.tight_layout()
-        self.convergence_canvas.draw()
-        
-        # Hide "no data" message
-        self.convergence_no_data_label.setVisible(False)
-        
-        # ---------- RELATIVE CHANGE PLOT ----------
-        # Clear the relative change figure
-        self.rel_change_fig.clear()
-        
-        # Check if we have relative change data
-        if len(results["relative_changes"]) > 1:
-            # Get the data for relative change plot
-            rel_changes = results["relative_changes"][1:]  # Skip first (which is NaN)
-            points = omega_points[1:]  # Skip first point to match rel_changes
-            threshold = self.sensitivity_threshold.value()
-            step_size = self.sensitivity_step_size.value()  # Get step size for annotations
-            
-            # Create the plot
-            ax = self.rel_change_fig.add_subplot(111)
-            
-            # Use log scale for better visibility of small changes
-            ax.semilogy(points, rel_changes, 'o-', linewidth=2.5, markersize=8, 
-                      color='#5D4954', alpha=0.8, label=f'Relative Change (Latest: {rel_changes[-1]:.6f})')
-            
-            # Add threshold line
-            ax.axhline(y=threshold, color='red', linestyle='--', linewidth=2, 
-                      label=f'Threshold: {threshold:.6f}')
-            
-            # Color points based on whether they're below threshold
-            for i, (p, rc) in enumerate(zip(points, rel_changes)):
-                if rc < threshold:
-                    ax.scatter(p, rc, s=120, color='green', zorder=5, alpha=0.7,
-                              marker='o', edgecolors='black')
-            
-            # Highlight the convergence point specifically
-            convergence_point = results.get("convergence_point")
-            if convergence_point is not None and convergence_point in points:
-                try:
-                    conv_idx = list(points).index(convergence_point)
-                    conv_y = rel_changes[conv_idx]
-                    ax.scatter(convergence_point, conv_y, s=220, color='green', zorder=7, 
-                              marker='*', label=f'First Convergence ({convergence_point} pts)')
-                    
-                    # Draw a vertical line at convergence point
-                    ax.axvline(x=convergence_point, color='green', linestyle=':', linewidth=1.5, alpha=0.7)
-                    
-                    # Add annotation explaining this is where convergence was first detected
-                    ax.annotate('First convergence point', 
-                               xy=(convergence_point, conv_y),
-                               xytext=(convergence_point + step_size/2, max(conv_y * 1.5, threshold * 2)),
-                               arrowprops=dict(facecolor='green', shrink=0.05, width=1.5, headwidth=8),
-                               fontsize=10, color='green', fontweight='bold')
-                except (ValueError, IndexError):
-                    pass
-            
-            # Find consecutive points below threshold for visualization
-            consecutive_points = []
-            for i in range(len(rel_changes) - 1):
-                if rel_changes[i] < threshold and rel_changes[i+1] < threshold:
-                    consecutive_points.append((points[i], rel_changes[i]))
-                    consecutive_points.append((points[i+1], rel_changes[i+1]))
-            
-            # Highlight all points below threshold
-            if consecutive_points:
-                # Extract unique points that are in consecutive pairs
-                unique_consecutive = list(set(consecutive_points))
-                x_cons = [p[0] for p in unique_consecutive]
-                y_cons = [p[1] for p in unique_consecutive]
-                
-                # Don't include convergence point to avoid duplicate marker
-                if convergence_point is not None:
-                    filtered_x = []
-                    filtered_y = []
-                    for i, x in enumerate(x_cons):
-                        if x != convergence_point:
-                            filtered_x.append(x)
-                            filtered_y.append(y_cons[i])
-                    
-                    if filtered_x:                      # Only plot if there are points to show
-                        ax.scatter(filtered_x, filtered_y, s=150, color='#00AA00', zorder=6, 
-                                  marker='o', label=f'Stable Points ({len(filtered_x)})')
-            
-            # Add grid and labels
-            ax.grid(True, linestyle='--', alpha=0.7)
-            ax.set_xlabel('Number of Omega Points', fontsize=14, fontweight='bold')
-            ax.set_ylabel('Relative Change (log scale)', fontsize=14, fontweight='bold')
-            ax.set_title('Relative Change in Slope with Increasing Frequency Resolution', 
-                        fontsize=16, fontweight='bold')
-            
-            # Create enhanced legend with better formatting
-            legend = ax.legend(fontsize=12, loc='best', frameon=True, fancybox=True, shadow=True,
-                             title="Relative Change Analysis Results", title_fontsize=14,
-                             borderpad=1.2, labelspacing=1.2, handletextpad=1.0)
-        else:
     def save_sobol_results(self):
         options = QFileDialog.Options()
         file_path, _ = QFileDialog.getSaveFileName(self, "Save Sobol Results", "",
@@ -3628,179 +1082,4 @@ class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin
         self.status_bar.showMessage("Reset to default values")
         # Reset logic for all parameters would be implemented here
 
-
-    def get_main_system_params(self):
-        """Get the main system parameters from the UI"""
-        return (
-            self.mu_box.value(),
-            *[b.value() for b in self.landa_boxes],
-            *[b.value() for b in self.nu_boxes],
-            self.a_low_box.value(),
-            self.a_up_box.value(),
-            self.f_1_box.value(),
-            self.f_2_box.value(),
-            self.omega_dc_box.value(),
-            self.zeta_dc_box.value()
-        )
-        
-    def update_sobol_plot(self):
-        """Update the Sobol analysis plot based on the selected plot type"""
-        key = self.sobol_combo.currentText()
-        if key in self.sobol_plots:
-            fig = self.sobol_plots[key]
-            self.sobol_canvas.figure = fig
-            self.sobol_canvas.draw()
-        else:
-            self.sobol_canvas.figure.clear()
-        self.sobol_canvas.draw()
-
-    def create_frequency_tab(self):
-        """Create the frequency and plot tab"""
-        self.freq_tab = QWidget()
-        layout = QVBoxLayout(self.freq_tab)
-
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Frequency range group
-        freq_group = QGroupBox("Frequency Range & Plot Options")
-        freq_layout = QFormLayout(freq_group)
-
-        # OMEGA start
-        self.omega_start_box = QDoubleSpinBox()
-        self.omega_start_box.setRange(0, 1e6)
-        self.omega_start_box.setDecimals(6)
-        self.omega_start_box.setValue(0.0)
-        freq_layout.addRow("Ω Start:", self.omega_start_box)
-
-        # OMEGA end
-        self.omega_end_box = QDoubleSpinBox()
-        self.omega_end_box.setRange(0, 1e6)
-        self.omega_end_box.setDecimals(6)
-        self.omega_end_box.setValue(10000.0)
-        freq_layout.addRow("Ω End:", self.omega_end_box)
-
-        # OMEGA points
-        self.omega_points_box = QSpinBox()
-        self.omega_points_box.setRange(1, 1000000000)  # Increased to 10^9
-        self.omega_points_box.setValue(1200)
-        freq_layout.addRow("Ω Points:", self.omega_points_box)
-
-        # Plot options
-        self.plot_figure_chk = QCheckBox("Plot Figure")
-        self.plot_figure_chk.setChecked(True)
-        freq_layout.addRow(self.plot_figure_chk)
-        
-        self.show_peaks_chk = QCheckBox("Show Peaks")
-        self.show_peaks_chk.setChecked(False)
-        freq_layout.addRow(self.show_peaks_chk)
-        
-        self.show_slopes_chk = QCheckBox("Show Slopes")
-        self.show_slopes_chk.setChecked(False)
-        freq_layout.addRow(self.show_slopes_chk)
-        
-        main_layout.addWidget(freq_group)
-        
-        # Add interpolation options section
-        interp_group = QGroupBox("Curve Interpolation Options")
-        interp_layout = QFormLayout(interp_group)
-        
-        # Interpolation method combo box
-        self.interp_method_combo = QComboBox()
-        from modules.FRF import INTERPOLATION_METHODS
-        self.interp_method_combo.addItems(['none'] + INTERPOLATION_METHODS)
-        self.interp_method_combo.setCurrentText('cubic')  # Default to cubic
-        self.interp_method_combo.setToolTip(
-            "none: No interpolation (raw data)\n"
-            "linear: Straight line segments\n"
-            "cubic: Smooth cubic spline (default)\n"
-            "quadratic: Quadratic interpolation\n"
-            "nearest: Nearest neighbor interpolation\n"
-            "akima: Reduced oscillation spline\n"
-            "pchip: Piecewise cubic Hermite\n"
-            "smoothing_spline: Smoothing spline\n"
-            "bspline: B-spline interpolation\n"
-            "savgol: Savitzky-Golay filter (smoothing)\n"
-            "moving_average: Moving average smoothing\n"
-            "gaussian: Gaussian filter smoothing\n"
-            "bessel: Bessel filter (good for frequency data)\n"
-            "barycentric: Barycentric interpolation\n"
-            "rbf: Radial basis function"
-        )
-        interp_layout.addRow("Interpolation Method:", self.interp_method_combo)
-        
-        # Points to use for interpolation
-        self.interp_points_box = QSpinBox()
-        self.interp_points_box.setRange(100, 10000)
-        self.interp_points_box.setValue(1000)
-        self.interp_points_box.setSingleStep(100)
-        self.interp_points_box.setToolTip("Number of points to use in the interpolated curve")
-        interp_layout.addRow("Interpolation Points:", self.interp_points_box)
-        
-        # Add info label about interpolation
-        info_label = QLabel(
-            "Interpolation smooths the frequency response curve for better visualization. "
-            "Different methods provide various levels of smoothing and can affect how peaks "
-            "and transitions appear. 'none' shows the raw calculation points."
-        )
-        info_label.setWordWrap(True)
-        info_label.setStyleSheet("color: #666; font-style: italic;")
-        interp_layout.addRow(info_label)
-        
-        # Add interpolation group to main layout
-        main_layout.addWidget(interp_group)
-
-        # Add Run FRF button
-        run_frf_container = QWidget()
-        run_frf_layout = QHBoxLayout(run_frf_container)
-        run_frf_layout.setContentsMargins(0, 20, 0, 0)  # Add some top margin
-        
-        # Add stretch to push button to center
-        run_frf_layout.addStretch()
-        
-        # Create and style the Run FRF button
-        self.run_frf_button = QPushButton("Run FRF Analysis")
-        self.run_frf_button.setObjectName("primary-button")
-        self.run_frf_button.setMinimumWidth(200)  # Make button wider
-        self.run_frf_button.setMinimumHeight(40)  # Make button taller
-        self.run_frf_button.clicked.connect(self.run_frf)
-        self.run_frf_button.setStyleSheet("""
-            QPushButton#primary-button {
-                background-color: #4B67F0;
-                color: white;
-                border: none;
-                border-radius: 4px;
-                padding: 8px 20px;
-                font-weight: bold;
-                font-size: 14px;
-            }
-            QPushButton#primary-button:hover {
-                background-color: #3B57E0;
-            }
-            QPushButton#primary-button:pressed {
-                background-color: #2B47D0;
-            }
-        """)
-        run_frf_layout.addWidget(self.run_frf_button)
-        
-        # Add stretch to push button to center
-        run_frf_layout.addStretch()
-        
-        main_layout.addWidget(run_frf_container)
-        
-        # Add comparative visualization options
-        self.create_comparative_visualization_options(main_layout)
-        
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
